### 2024-04-24 #6675 @editorialbot
[PRE REVIEW]: ChainoPy: A Python Library for Discrete Time Markov Chain based stochastic analysis
**Submitting author:** @aadya940 (<a href="http://orcid.org/0009-0006-2794-7861">Aadya  Chinubhai</a>)
**Repository:** https://github.com/aadya940/chainopy
**Branch with paper.md** (empty if default branch): master
**Version:** v1.0.0
