### 2024-04-24 #6675 @crvernon
@aadya940 - in the meantime, could you work on fixing the author affiliation issue here:

> ⚠️ [An error happened when generating the pdf](https://github.com/openjournals/joss-papers/actions/runs/8816842321?check_suite_focus=true). Author (Aadya A. Chinubhai) is missing affiliation.

You will also need to reduce the length of your paper to the JOSS requirement of being between 250-1000 words (see https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain).

Thank you!
