### 2024-04-26 #6675 @mooniean
Hi @aadya940! I'll be editing your paper!

Have you addressed the issues mentioned by Chris? If so, if we do
`@editorialbot generate pdf` we can see the new version of the pdf.

Do you have any reviewer suggestions, before I contact anyone?
