+++ JOSS/paper.md
@@ -143,7 +143,7 @@ The results are as follows:
 |                        | 5000    | 209 ms        | 16.4 ms          | 285 ms      | 27.6ms         |
 
 Apart from this, we test the `MarkovChainNeuralNetworks` by training them and comparing random walks between
-the original `MarkovChain` (Right) object and those generated by `MarkovChainNeuralNetworks` (Left) through a Histogram.
+the original `MarkovChain` object and those generated by `MarkovChainNeuralNetworks` through a Histogram.
 
 ## Conclusion
 
