### 2024-05-15 #6675 @tbrown122387
> Hello 👋  @tbrown122387 @thomaspinder and @braniii - Would you be willing to review this submission to JOSS?
>
>
>
> > https://github.com/openjournals/joss-reviews/issues/6675
>
>
>
> We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

I'm too busy. Sorry!
