+++ README.md
@@ -30,7 +30,7 @@ cd chainopy
 ### 2. Install Dependencies
 ```bash
 pip install -r requirements.txt
-pip install -r requirements-test.txt
+pip install -r requirements_test.txt
 ```
 
 ### 3. Compile Cython Modules
