+++ chainopy/markov_chain.py
@@ -44,6 +44,12 @@ class MarkovChain:
         self._validate_transition_matrix(p, states, epsilon=1e-16)
 
     def __repr__(self) -> str:
+        if self.tpm is None:
+            return (
+            f"<Object of type MarkovChain with uninitialized transition matrix  "
+            f"\n"
+            f"and unknown states>"
+        )
         return (
             f"<Object of type MarkovChain with {self.tpm.shape[0]} x {self.tpm.shape[1]} sized transition matrix  "
             f"\n"
