### 2024-05-17 6dfcb7e
minor modification __str__
