### 2024-06-07 #6675 @mooniean
Hello 👋 @MichalisPanayides @RomainAzais - Would you be willing to review this submission to JOSS?

> https://github.com/openjournals/joss-reviews/issues/6675
