### 2024-06-11 #6675 @MichalisPanayides
> Hello 👋 @MichalisPanayides @RomainAzais - Would you be willing to review this submission to JOSS?
>
> > #6675
>
> We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Hi there 👋 I'd be happy to review this 😃
