### 2024-06-12 #6880 @mooniean
Hi @aadya940  and reviewers @braniii @MichalisPanayides 👋 ! - this is the review thread for the submission. All of our communications will happen here from now on.

Meanwhile, please check the post at the top of the issue for instructions on how to generate your own review checklist. As you go over the submission, please check any items that you feel have been satisfied. There are also links to the JOSS reviewer guidelines.
