### 2024-06-12 #6675 @RomainAzais
Hello,
Sorry, but I am too busy at the moment.
Best

> Hello 👋 @MichalisPanayides @RomainAzais - Would you be willing to review this submission to JOSS?
>
> > #6675
>
> We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
