### 2024-06-18 f138763
Minor bug fixes in README.md
