### 2024-06-18 aadya940/chainopy#4 @MichalisPanayides
JOSS submission review
This issue is part of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/6880). Here's a list of some comments:

- The installation process listed on the `README.md` along with the tests worked fine
- There's numerous bugs in [The Basics](https://github.com/aadya940/chainopy?tab=readme-ov-file#the-basics) section of the `README.md` file. All the suggested changes have been outlined in this [PR](https://github.com/aadya940/chainopy/pull/3)
    - Would be nice to have a bit more text that describes what is happening in that section
    - Also possibly add a graph there similar to the one in the [final example](https://github.com/aadya940/chainopy?tab=readme-ov-file#example---apple-weekly-high-stock-data-prediction-using-chainopymarkovswitchingmodel)
- In the `README.md` file it would also be nice to mention the jupyter notebooks in `examples/` and `benchmark/` directories
   - It would be also nice to mention that some additional imports are required to run these (e.g. `seaborn`, `yfinance`)
- Would also be great to have a `How to contribute` section
- There seem to be some bugs when attempting to run the jupyter notebooks. e.g:
    - In `examples/demo.ipynb` the line `sim = x.simulate(initial_state="Rainy", n_steps=500)` seems to get the following error (which seems to be fixed if you run `sim = x.simulate("Rainy", 500)` instead). I believe there is a bug in the `exceptions.py` file:
    ```python
    [38](https://file+.vscode-resource.vscode-cdn.net/Users/Desktop/code/chainopy/examples/~/anaconda3/envs/joss-chainopy/lib/python3.11/site-packages/chainopy/exceptions.py:38) elif func.__name__ == "simulate":
    ---> [39](https://file+.vscode-resource.vscode-cdn.net/Users/Desktop/code/chainopy/examples/~/anaconda3/envs/joss-chainopy/lib/python3.11/site-packages/chainopy/exceptions.py:39)     n_steps = args[2]
         [40](https://file+.vscode-resource.vscode-cdn.net/Users/Desktop/code/chainopy/examples/~/anaconda3/envs/joss-chainopy/lib/python3.11/site-packages/chainopy/exceptions.py:40)     initial_state = args[1]
         [41](https://file+.vscode-resource.vscode-cdn.net/Users/Desktop/code/chainopy/examples/~/anaconda3/envs/joss-chainopy/lib/python3.11/site-packages/chainopy/exceptions.py:41)     if not isinstance(n_steps, int) or n_steps <= 0:

    IndexError: tuple index out of range
    ```
    - In `examples/demo.ipynb` the line `new_object1.load_model("first-save.json")` also gets an error
    - In `benchmarks/benchmarks_fit.ipynb` the line `chainopy_mc.fit(words)`, under `Length = 50` also produces an error:
    ```python
    ---> [95](https://file+.vscode-resource.vscode-cdn.net/Users/Desktop/code/chainopy/benchmarks/~/anaconda3/envs/joss-chainopy/lib/python3.11/site-packages/chainopy/markov_chain.py:95)     raise ValueError(f"Invalid TPM {p}")
         [97](https://file+.vscode-resource.vscode-cdn.net/Users/Desktop/code/chainopy/benchmarks/~/anaconda3/envs/joss-chainopy/lib/python3.11/site-packages/chainopy/markov_chain.py:97) if not np.allclose(np.sum(p, axis=1), 1, atol=epsilon * len(states)):
         [98](https://file+.vscode-resource.vscode-cdn.net/Users/Desktop/code/chainopy/benchmarks/~/anaconda3/envs/joss-chainopy/lib/python3.11/site-packages/chainopy/markov_chain.py:98)     raise ValueError(
         [99](https://file+.vscode-resource.vscode-cdn.net/Users/Desktop/code/chainopy/benchmarks/~/anaconda3/envs/joss-chainopy/lib/python3.11/site-packages/chainopy/markov_chain.py:99)         f"Rows of the Transition Probability Matrix \
        [100](https://file+.vscode-resource.vscode-cdn.net/Users/Desktop/code/chainopy/benchmarks/~/anaconda3/envs/joss-chainopy/lib/python3.11/site-packages/chainopy/markov_chain.py:100)                             (TPM) {p} must sum to 1."
        [101](https://file+.vscode-resource.vscode-cdn.net/Users/Desktop/code/chainopy/benchmarks/~/anaconda3/envs/joss-chainopy/lib/python3.11/site-packages/chainopy/markov_chain.py:101)     )

    ValueError: Invalid TPM [[1.e-16 1.e-16 1.e-16 ... 1.e-16 1.e-16 1.e-16]
     [1.e-16 1.e-16 1.e-16 ... 1.e-16 1.e-16 1.e-16]
     [1.e-16 1.e-16 1.e-16 ... 1.e-16 1.e-16 1.e-16]
     ...
     [1.e+00 1.e-16 1.e-16 ... 1.e-16 1.e-16 1.e-16]
     [1.e-16 1.e-16 1.e-16 ... 1.e-16 1.e-16 1.e-16]
     [1.e-16 1.e-16 1.e-16 ... 1.e-16 1.e-16 1.e-16]]
    ```
