### 2024-06-18 d841ab7
bug fix in exceptions.py
