### 2024-06-18 9e92e90
fix bug in .fit
