### 2024-06-18 da2f8ee
fix load_model
