### 2024-06-18 0273576
add extra requirements for demo
