+++ examples/requirements_demo_extra.txt
@@ -1,0 +1,2 @@
+seaborn
+yfinance
