### 2024-06-23 aadya940/chainopy#4 @MichalisPanayides
@aadya940 Thanks for addressing all my comments. I can confirm that the contents of all notebooks run fine now. Here's a couple more comments:

- Running the notebooks in the `benchmarks/` directory I found that the `fit` method seems to not satisfy the claims presented in the [paper](https://raw.githubusercontent.com/openjournals/joss-papers/joss.06880/joss.06880/10.21105.joss.06880.pdf). More precisely when running the `benchmarks_fit.ipynb` notebook it seems like for `Length = 500`, `Length = 1000`, `Length = 5000` and `Length = 10000` the `fit` mehod of `chanopy` seems to be slower than the `fit_sequence` method of pydtmc. Please let me know if this makes sense and/or if I have misunderstood something. (Picture of the results of `Length = 500` running in a local jupyter notebook)

![image](https://github.com/aadya940/chainopy/assets/20097997/c1dd9988-1aea-40ed-9674-e210d235906c)

- Regarding the software paper I think some of the tables might need some editing. I am slightly confused about the alignment of the `Number of Words`/`Transition-Matrix size` row and the `Mean`/`St. dev.` columns. Apologies if I'm missing something there. This comment relates to the following tables:
        - `is_absorbing` Methods
        - `stationary_dist` vs `pi` Methods
        - `fit` vs `fit_sequence` Method
