### 2024-06-24 aadya940/chainopy#4 @aadya940
Hello, it was quite strange to know this happened in the `fit` function. However, upon careful inspection I realized the caching decorator was missing in the `fit` method for some reason which was present earlier while I was writing benchmarks in the paper. Could just be a silly mistake. I assume it will work fine now and am attaching a screenshot below.

![image](https://github.com/aadya940/chainopy/assets/77720426/17744743-8ef6-4e4a-8d7c-951f9ae27cd4)
