### 2024-06-24 1703477
Fix MarkovSwitchingModel breakage
