### 2024-06-24 aadya940/chainopy#4 @aadya940
Merged the PR.

Regarding the paper, I see the issue, I'll rewrite the tables a bit.
