### 2024-06-26 aadya940/chainopy#17 @braniii
[JOSS Review] Functionality section
Hi @aadya940,

this is part of [my review](https://github.com/openjournals/joss-reviews/issues/6880) and will be updated. I will try to work through the review list in the next weeks.

- [x] publish to PyPI or conda-forge: In the [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html#installation-instructions) it is clearly mentioned, that Python packages should be `pip install`able. Having setup `setup.py`  it should be easy to publish this package to pypi or conda-forge, or to make it at least installable from github via `pip install git+https://github.com/aadya940/chainopy`. This should also make it more likely that others will try out (and therefor cite) this package. For your optional deps, you could use `extras_require` in the setup.
- [x] In the installation instruction (README) the Github link of this repository has a copy+paste error `username` instead of your username
- [x] In `requirements.txt` cython is listed twice.
- [x] Numpy is listed in both, `requirements.txt` and `requirements_test.txt`
- [x] Fix installation via PyPI (see below), so include binaries for all supported python versions
- [x] Include dependencies as requirements
- [x] ~~Include optional dependencies [docs/tests], e.g. as `extra_require`, to allow for easy setup.~~ I saw you ticked it without including. But I guess this is no requirement and therefore, I remove this point.
- [x] What are you using numba for? I cannot spot any usage in the code.
- [x] Fix import issue (see below)

Daniel Nagel
