### 2024-07-02 aadya940/chainopy#18 @braniii
[JOSS Review] Documentation
Hi @aadya940,

this is part of [my review](https://github.com/openjournals/joss-reviews/issues/6880) and will be updated the next days.

Here is a list with open tasks:

- [x] Host the documentation
- [x] Fix sphinx so that the API is included in the documentation
- [x] Add examples for the advertised functionality in the docs.
- [x] Would it be possible to include the API of the public methods, rather than of the private submodules?
- [x] Please add a short description of the 3 classes, including a short description of the attributes. Only a name without a type, e.g. `epsilon` is (at least for me) not easy to interpret.
- [x] Fix rendering issues of the README. in "the basics" section.

I've finally finished the review. If you have any questions, feel free to write me.

Kind regards
Daniel Nagel
