+++ .readthedocs.yaml
@@ -17,12 +17,16 @@ python:
         - build
         - docs
 
-before_build:
-  jobs:
-    - 'sphinx-apidoc -o docs/source chainopy/'
-
+build:
+  before_build:
+    - python setup.py build_ext --inplace
+    - sphinx-apidoc -o docs/source chainopy/
 
 sphinx:
   builder: html
   configuration: docs/source/conf.py
   fail_on_warning: false
+
+submodules:
+  include: true
+  recursive: true
