+++ .readthedocs.yaml
@@ -3,7 +3,6 @@ version: 2
 formats: all
 
 build:
-  os: ubuntu-22.04
   tools:
     python: "3.10"
 
