+++ .readthedocs.yaml
@@ -3,6 +3,7 @@ version: 2
 formats: all
 
 build:
+  os: ubuntu-20.04
   tools:
     python: "3.10"
 
