+++ .readthedocs.yaml
@@ -2,11 +2,13 @@ version: 2
 
 formats: all
 
-os: ubuntu-20.04
-
 build:
+  os: ubuntu-20.04
   tools:
     python: "3.10"
+  before_build:
+    - python setup.py build_ext --inplace
+    - sphinx-apidoc -o docs/source chainopy/
 
 python:
   install:
@@ -17,11 +19,6 @@ python:
       extra_requirements:
         - build
         - docs
-
-build:
-  before_build:
-    - python setup.py build_ext --inplace
-    - sphinx-apidoc -o docs/source chainopy/
 
 sphinx:
   builder: html
