+++ .readthedocs.yaml
@@ -24,7 +24,3 @@ sphinx:
   builder: html
   configuration: docs/source/conf.py
   fail_on_warning: false
-
-submodules:
-  include: true
-  recursive: true
