+++ docs/source/conf.py
@@ -26,9 +26,14 @@ extensions = [
     "sphinx.ext.autodoc",
 ]
 
-templates_path = ["_templates"]
-exclude_patterns = []
+source_suffix = {
+    ".md": "markdown",
+}
 
+templates_path = ["_templates"]
+exclude_patterns = [
+    ".ipynb_checkpoints",
+]
 
 # -- Options for HTML output -------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#options-for-html-output
