+++ docs/source/conf.py
@@ -26,12 +26,9 @@ extensions = [
     "sphinx.ext.autodoc",
 ]
 
-source_suffix = {
-    ".md": "markdown",
-}
-
 templates_path = ["_templates"]
 exclude_patterns = [
+    "build",
     ".ipynb_checkpoints",
 ]
 
