+++ docs/source/conf.py
@@ -26,11 +26,22 @@ extensions = [
     "sphinx.ext.autodoc",
 ]
 
+source_suffix = {
+    ".rst": "restructuredtext",
+    ".md": "markdown",
+}
+
 templates_path = ["_templates"]
 exclude_patterns = [
     "build",
     ".ipynb_checkpoints",
 ]
+
+autodoc_default_options = {
+    "members": True,
+    "inherited-members": True,
+    "member-order": "bysource",
+}
 
 # -- Options for HTML output -------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#options-for-html-output
