+++ .readthedocs.yaml
@@ -10,6 +10,7 @@ build:
   jobs:
     pre_build:
       - python setup.py build_ext --inplace
+  jobs:
     - sphinx-apidoc -o docs/source chainopy/
 
 python:
