+++ .readthedocs.yaml
@@ -29,5 +29,5 @@ sphinx:
   fail_on_warning: false
 
 submodules:
-  include: []
+  include: ['chainopy']
   recursive: true
