+++ .readthedocs.yaml
@@ -11,7 +11,7 @@ build:
     pre_build:
       - python setup.py build_ext --inplace
   jobs:
-    - sphinx-apidoc -o docs/source chainopy/
+    - sphinx-apidoc -o docs/source chainopy/ *.pyx
 
 python:
   install:
