+++ .readthedocs.yaml
@@ -8,7 +8,8 @@ build:
   tools:
     python: "3.10"
   jobs:
-    - sphinx-apidoc -o docs/source chainopy/ *.pyx
+    sphinx-apidoc:
+      script: sphinx-apidoc -o docs/source chainopy/ *.pyx
 
 python:
   install:
