+++ docs/source/conf.py
@@ -10,7 +10,7 @@ import os
 import sys
 
 # Root Path
-sys.path.insert(0, os.path.abspath("../../."))
+sys.path.insert(0, os.path.abspath("../../chainopy"))
 
 
 project = "Chainopy"
@@ -24,6 +24,18 @@ release = "1.0"
 extensions = [
     "myst_parser",
     "sphinx.ext.autodoc",
+    'sphinx.ext.napoleon',
+    'sphinx.ext.viewcode',
+    'sphinx.ext.autosummary',
+    'sphinx.ext.doctest',
+    'sphinx.ext.coverage',
+    'sphinx.ext.intersphinx',
+    'sphinx.ext.todo',
+    'sphinx.ext.mathjax',
+    'sphinx.ext.ifconfig',
+    'sphinx.ext.githubpages',
+    'sphinx.ext.extlinks',
+    'sphinx_autodoc_typehints',
 ]
 
 bibtex_bibfiles = ["paper.bib"] 
@@ -43,9 +55,30 @@ exclude_patterns = [
 ]
 
 autodoc_default_options = {
-    "members": True,
-    "inherited-members": True,
-    "member-order": "bysource",
+    'members': True,
+    'undoc-members': True,
+    'private-members': True,
+    'special-members': '__init__',
+    'inherited-members': True,
+    'show-inheritance': True,
+}
+
+napoleon_google_docstring = True
+napoleon_numpy_docstring = True
+napoleon_include_init_with_doc = True
+napoleon_include_private_with_doc = False
+napoleon_include_special_with_doc = True
+napoleon_use_admonition_for_examples = False
+napoleon_use_admonition_for_notes = False
+napoleon_use_admonition_for_references = False
+napoleon_use_ivar = False
+napoleon_use_param = True
+napoleon_use_rtype = True
+
+intersphinx_mapping = {
+    'python': ('https://docs.python.org/3', None),
+    'numpy': ('https://numpy.org/doc/stable/', None),
+    'scipy': ('https://docs.scipy.org/doc/scipy/reference/', None),
 }
 
 # -- Options for HTML output -------------------------------------------------
@@ -53,3 +86,13 @@ autodoc_default_options = {
 
 html_theme = "furo"
 html_static_path = ["_static"]
+
+def skip_cython_docstring(app, what, name, obj, skip, options):
+    # Skip docstrings in Cython files
+    if what == "module" and name.endswith(".pyx"):
+        return True
+    return skip
+
+def setup(app):
+    # Call event.
+    app.connect('autodoc-skip-member', skip_cython_docstring)
