+++ requirements_doc.txt
@@ -2,3 +2,4 @@ sphinx
 sphinx-rtd-theme
 furo
 myst-parser
+sphinx-autodoc-typehints
