+++ .readthedocs.yaml
@@ -1,8 +1,5 @@
 version: 2
 
-formats: 
-  - htmlzip
-
 build:
   os: ubuntu-20.04
   tools:
