+++ docs/source/index.md
@@ -12,11 +12,6 @@
 
 # Indices and tables
 
-- {ref}`genindex`
-- {ref}`modindex`
-- {ref}`search`
-
-
 ```{include} ../../README.md
 ```
 
