+++ .readthedocs.yaml
@@ -5,7 +5,7 @@ build:
   tools:
     python: "3.10"
   jobs:
-    post_build:
+    pre_build:
       - sphinx-apidoc -o docs/source chainopy/ *.pyx
 
 python:
