### 2024-07-10 aadya940/chainopy#18 @braniii
Here is a list with minor issues I found while reviewing this repository. I thought it would be helpful sharing them:

- [ ] in `test_markovchain.py` line 73+74 and 88+89 are without functionality. Line 75 should be `if not (fm_x is None and fm_y is None)`:, but having a deterministic input, I guess line 75+76 are never reached and should be deleted!? The same for line 88+89. If you want to run the same test with different input parameters you can use `@pytest.mark.parametrize` , see https://docs.pytest.org/en/6.2.x/parametrize.html
