### 2024-07-10 aadya940/chainopy#18 @aadya940
Yes @braniii , you are correct regarding the `test_markovchain.py`, I'll fix it soon. As for the docs, there is an issue with readthedocs, they give a 404 error which I'm unable to solve however you'll be able to build the docs locally. Also, the docs are essentially generated using the docstring so you could just refer to those as of now.

Here are the steps to generate the docs locally:

Go to the project folder:

Step - 1:

```bash
pip install -r requirements_doc.txt
```

Step - 2:

```bash
sphinx-apidoc -o docs/source chainopy/
```

Step - 3:
```bash
cd docs
make html
```

Step - 4:
```bash
cd build/html
python -m http.server
```

The docs will be listed on the given port.
