### 2024-07-10 aadya940/chainopy#18 @aadya940
As for uploading the package on `pip`, I'll do it ASAP, sorry for the delay :))
