### 2024-07-10 aadya940/chainopy#18 @braniii
> Yes @braniii , you are correct regarding the `test_markovchain.py`, I'll fix it soon. As for the docs, there is an issue with readthedocs, they give a 404 error which I'm unable to solve however you'll be able to build the docs locally. Also, the docs are essentially generated using the docstring so you could just refer to those as of now.
>
> Here are the steps to generate the docs locally:
>
> Go to the project folder:
>
> Step - 1:
>
> ```shell
> pip install -r requirements_doc.txt
> ```
>
> Step - 2:
>
> ```shell
> sphinx-apidoc -o docs/source chainopy/
> ```
>
> Step - 3:
>
> ```shell
> cd docs
> make html
> ```
>
> Step - 4:
>
> ```shell
> cd build/html
> python -m http.server
> ```
>
> The docs will be listed on the given port.

I've managed to build the docs, but it seems that there are some markdown parsing issues. I will wait until the documentation is back online.

![Screenshot from 2024-07-10 16-06-12](https://github.com/aadya940/chainopy/assets/28560129/0e6a610f-84df-427c-a3c2-22c7682a605c)
![Screenshot from 2024-07-10 16-06-55](https://github.com/aadya940/chainopy/assets/28560129/cf00216a-46ba-4eb5-a815-faf3ea433f3f)
