+++ chainopy/markov_chain.py
@@ -105,7 +105,7 @@ class MarkovChain:
         It will override the current transition-matrix.
 
         Args:
-        ----
+        -----
             data: Union[str, list]
                 Data on which the MarkovChain model must
                 be fitted.
@@ -113,11 +113,11 @@ class MarkovChain:
                 Small dummy value to avoid zeros in the Transition-Matrix
 
         Returns:
-        -------
+        --------
             ndarray: Transition - Matrix based on `data`
 
         Usage:
-        -----
+        ------
             >>> chainopy.MarkovChain().fit("My name is John.")
         """
         return self._learn_matrix(data=data, epsilon=epsilon)
@@ -136,14 +136,14 @@ class MarkovChain:
         Simulate the Markov Chain for `n_steps` steps.
 
         Args:
-        ----
+        -----
             initial_state: str
                 State from which the simulation starts
             n_steps: int
                 Number of steps to simulate the chain for
 
         Returns:
-        -------
+        --------
             list: Contains states attained during simulation
         """
         return _simulate._simulate_cython(self.states, self.tpm, initial_state, n_steps)
@@ -162,7 +162,7 @@ class MarkovChain:
     def adjacency_matrix(self) -> np.ndarray:
         """
         Returns:
-        -------
+        --------
             ndarray: Adjacency matrix of the chain.
         """
         return (self.tpm > self.epsilon).astype(int)
@@ -172,12 +172,12 @@ class MarkovChain:
         Return the next most likely states.
 
         Args:
-        ----
+        -----
             initial_state : str
                 Initial state.
 
         Returns:
-        -------
+        --------
             str: Next most likely state.
         """
         initial_vect = self._vectorize(self.states, initial_state)
@@ -189,12 +189,12 @@ class MarkovChain:
         Calculates the distribution of the Markov Chain after n-steps.
 
         Args:
-        ----
+        -----
             n_steps : int
                 Number of steps.
 
         Returns:
-        -------
+        --------
             ndarray: Distribution of the Markov Chain.
         """
         is_eigendecom = self.eigendecom
@@ -217,7 +217,7 @@ class MarkovChain:
         Checks if the Markov chain is ergodic.
 
         Returns:
-        -------
+        --------
             bool: True if the Markov chain is ergodic, False otherwise.
         """
         return self.is_irreducible() and self.is_aperiodic()
@@ -228,7 +228,7 @@ class MarkovChain:
         Checks if the Markov chain is symmetric.
 
         Returns:
-        -------
+        --------
             bool: True if the Markov chain is symmetric, False otherwise.
         """
         return np.allclose(self.tpm, self.tpm.transpose())
@@ -239,7 +239,7 @@ class MarkovChain:
         Returns the stationary distribution of the Markov chain.
 
         Returns:
-        -------
+        --------
             ndarray: Stationary distribution.
         """
         tpm_T = self.tpm.transpose()
@@ -252,7 +252,7 @@ class MarkovChain:
         Checks if two states are communicating.
 
         Args:
-        ----
+        -----
             state1 : str
                 First state.
             state2 : str
@@ -261,7 +261,7 @@ class MarkovChain:
                 Threshold for convergence. Defaults to 1000.
 
         Returns:
-        -------
+        --------
             bool: True if the states are communicating, False otherwise.
         """
         return _is_communicating.is_communicating_cython(
@@ -274,7 +274,7 @@ class MarkovChain:
         Checks if the Markov chain is irreducible.
 
         Returns:
-        -------
+        --------
             bool: True if the Markov chain is irreducible, False otherwise.
         """
         return all(
@@ -291,7 +291,7 @@ class MarkovChain:
         Returns all absorbing states.
 
         Returns:
-        -------
+        --------
[…]
