+++ docs/source/index.rst
@@ -1,0 +1,19 @@
+# Welcome to Chainopy's documentation!
+
+CHAINOPY
+========
+
+.. toctree::
+
+.. autosummary::
+   :toctree: _autosummary
+
+   chainopy._markov_chain
+   chainopy._markov_switching
+   chainopy._nn
+
+
+README File
+===========
+
+.. mdinclude:: ../../README.md
