+++ docs/source/index.md
@@ -1,21 +1,0 @@
-% Chainopy documentation master file, created by
-% sphinx-quickstart on Sat Feb 10 13:29:00 2024.
-% You can adapt this file completely to your liking, but it should at least
-% contain the root `toctree` directive.
-
-# Welcome to Chainopy's documentation!
-
-```{toctree}
-:caption: 'Contents:'
-:maxdepth: 2
-```
-
-# Indices and tables
-
-- {ref}`genindex`
-- {ref}`modindex`
-- {ref}`search`
-
-
-```{include} ../../README.md
-```
