+++ .github/workflows/ci.yml
@@ -75,7 +75,7 @@ jobs:
         run: pip install -r requirements_doc.txt
 
       - name: Run apidoc
-        run: sphinx-apidoc -o docs/source chainopy/
+        run: sphinx-apidoc --private -o docs/source ../chainopy/
 
       - name: Build HTML documentation
         run: sphinx-build -b html docs/source docs/build/html
