+++ .readthedocs.yaml
@@ -7,7 +7,7 @@ build:
   jobs:
     pre_build:
       - python setup.py build_ext --inplace
-      - sphinx-apidoc --private -o docs/source ../chainopy/
+      - sphinx-apidoc --private -o docs/source chainopy/
 
 python:
   install:
