+++ .github/workflows/ci.yml
@@ -75,7 +75,7 @@ jobs:
         run: pip install -r requirements_doc.txt
 
       - name: Run apidoc
-        run: sphinx-apidoc --private -o docs/source ../
+        run: sphinx-apidoc --private -o docs/source .
 
       - name: Build HTML documentation
         run: sphinx-build -b html docs/source docs/build/html
