+++ requirements.txt
@@ -1,5 +1,4 @@
 cython
-Cython
 numpy
 scipy
 xarray
