+++ docs/source/conf.py
@@ -53,6 +53,7 @@ exclude_patterns = [
     "chainopy._backend.rst",
     "chainopy.rst",
     "modules.rst",
+    "setup.rst",
     "_templates",
 ]
 
