+++ requirements.txt
@@ -1,3 +1,4 @@
+Cython
 cython
 numpy
 scipy
