+++ .github/workflows/ci.yml
@@ -74,6 +74,9 @@ jobs:
       - name: Install Doc dependencies
         run: pip install -r requirements_doc.txt
 
+      - name: Install build dependencies
+        run: pip install -r requirements.txt
+
       - name: Build Package
         run: python setup.py build_ext --inplace
 
