+++ chainopy/_markov_chain.py
@@ -104,20 +104,20 @@ class MarkovChain:
         Each Unique Word is considered a State.
         It will override the current transition-matrix.
 
-        Args:
-        -----
+        Args
+        ----
             data: Union[str, list]
                 Data on which the MarkovChain model must
                 be fitted.
             epsilon: float
                 Small dummy value to avoid zeros in the Transition-Matrix
 
-        Returns:
-        --------
+        Returns
+        -------
             ndarray: Transition - Matrix based on `data`
 
-        Usage:
-        ------
+        Usage
+        -----
             >>> chainopy.MarkovChain().fit("My name is John.")
         """
         return self._learn_matrix(data=data, epsilon=epsilon)
@@ -135,15 +135,15 @@ class MarkovChain:
         """
         Simulate the Markov Chain for `n_steps` steps.
 
-        Args:
-        -----
+        Args
+        ----
             initial_state: str
                 State from which the simulation starts
             n_steps: int
                 Number of steps to simulate the chain for
 
-        Returns:
-        --------
+        Returns
+        -------
             list: Contains states attained during simulation
         """
         return _simulate._simulate_cython(self.states, self.tpm, initial_state, n_steps)
@@ -161,8 +161,8 @@ class MarkovChain:
 
     def adjacency_matrix(self) -> np.ndarray:
         """
-        Returns:
-        --------
+        Returns
+        -------
             ndarray: Adjacency matrix of the chain.
         """
         return (self.tpm > self.epsilon).astype(int)
@@ -171,13 +171,13 @@ class MarkovChain:
         """
         Return the next most likely states.
 
-        Args:
-        -----
+        Args
+        ----
             initial_state : str
                 Initial state.
 
-        Returns:
-        --------
+        Returns
+        -------
             str: Next most likely state.
         """
         initial_vect = self._vectorize(self.states, initial_state)
@@ -188,13 +188,13 @@ class MarkovChain:
         """
         Calculates the distribution of the Markov Chain after n-steps.
 
-        Args:
-        -----
+        Args
+        ----
             n_steps : int
                 Number of steps.
 
-        Returns:
-        --------
+        Returns
+        -------
             ndarray: Distribution of the Markov Chain.
         """
         is_eigendecom = self.eigendecom
@@ -216,8 +216,8 @@ class MarkovChain:
         """
         Checks if the Markov chain is ergodic.
 
-        Returns:
-        --------
+        Returns
+        -------
             bool: True if the Markov chain is ergodic, False otherwise.
         """
         return self.is_irreducible() and self.is_aperiodic()
@@ -227,8 +227,8 @@ class MarkovChain:
         """
         Checks if the Markov chain is symmetric.
 
-        Returns:
-        --------
+        Returns
+        -------
             bool: True if the Markov chain is symmetric, False otherwise.
         """
         return np.allclose(self.tpm, self.tpm.transpose())
@@ -238,8 +238,8 @@ class MarkovChain:
         """
         Returns the stationary distribution of the Markov chain.
 
-        Returns:
-        --------
+        Returns
+        -------
             ndarray: Stationary distribution.
         """
         tpm_T = self.tpm.transpose()
@@ -251,8 +251,8 @@ class MarkovChain:
         """
         Checks if two states are communicating.
 
-        Args:
-        -----
+        Args
+        ----
             state1 : str
                 First state.
             state2 : str
@@ -260,8 +260,8 @@ class MarkovChain:
             threshold : int, optional
                 Threshold for convergence. Defaults to 1000.
 
-        Returns:
-        --------
+        Returns
+        -------
             bool: True if the states are communicating, False otherwise.
         """
         return _is_communicating.is_communicating_cython(
[…]
