### 2024-07-13 aadya940/chainopy#18 @aadya940
Hi @braniii , So the docs are up and running now at :
[https://chainopy.readthedocs.io/en/latest/](https://chainopy.readthedocs.io/en/latest/)

However, I'll improve these more in the future, this is just a start so that our reviews don't get affected. :))
