### 2024-07-13 aadya940/chainopy#17 @aadya940
Hey, I've uploaded chainopy on pip:

https://pypi.org/project/chainopy/1.0/

NOTE :
Due to the usage of Cython, the wheel is only compatible for linux platform at the moment, I'll compile and upload a windows wheel soon. Hope this works right noww
