+++ docs/source/conf.py
@@ -60,7 +60,6 @@ exclude_patterns = [
 autodoc_default_options = {
     "members": True,
     "undoc-members": True,
-    "private-members": True,
     "special-members": "__init__",
 }
 
