### 2024-07-14 aadya940/chainopy#4 @aadya940
@MichalisPanayides Do the tables look better now?
https://github.com/aadya940/chainopy/blob/master/JOSS/paper.md
