### 2024-07-14 318044d
parametrize test_fundamental_matrix
