### 2024-07-15 aadya940/chainopy#17 @braniii
Hi, I've tried out to install chainopy via PyPI (under linux) and was not successful. Here a short example to reproduce the error.

``` bash
# setting up a conda env
$ conda create -n chainopy python
$ conda activate chainopy

# install chainopy
$ python -m pip install chainopy
ERROR: Could not find a version that satisfies the requirement chainopy (from versions: none)
ERROR: No matching distribution found for chainopy
```
