### 2024-07-15 aadya940/chainopy#20 @braniii
[JOSS Review] Software Paper
Hi @aadya940,

this is the final part of [my review](https://github.com/openjournals/joss-reviews/issues/6880) and is about the paper.

Here is a list with open tasks:

- [x] Add a summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience.
- [x] Please improve on the summary. You are not following the typical structure of a scientific abstract.
- [x] You start the statement of need with the sentence "There are significant limitations in current Markov Chain packages that rely solely on NumPy (Harris et al., 2020) and Python for implementation." I guess, this highly depends on the field you are working on. In case of biophysics (Markov Modeling) most packages, e.g. pyemma, deeptime, msmhelper, are relying on Cython or Numba. So, I guess it would be helpful to cite the major packages of your community?
- [x] And regarding the Statement of Need. As far as I understand it, the necessity of chainopy is that it is faster than pyDTMC. When I look at the benchmarks, there are significant improvements in fast functions (<1ns) which probably do not limit existing workflows due to their fast execution. For the slower functions (simulate), on the other hand, there is a maximum of up to 50% performance improvement. Can you please comment on the extent to which ChainoPy allows new workflows, or whether the added value does not lie more in the additional algorithms.

Kind regards
Daniel Nagel
