+++ .github/workflows/release.yml
@@ -1,0 +1,152 @@
+name: Build and Publish Wheels
+
+on:
+  release:
+    types: [published]
+
+jobs:
+  build_wheels_linux:
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        python-version: [3.7, 3.8, 3.9, 3.10, 3.11, 3.12]
+        manylinux: [manylinux2010, manylinux2014]
+
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v2
+
+      - name: Set up Python ${{ matrix.python-version }}
+        uses: actions/setup-python@v2
+        with:
+          python-version: ${{ matrix.python-version }}
+
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install -r requirements.txt
+          pip install -r requirements_dev.txt
+          pip install -r requirements_doc.txt
+          pip install -r requirements_test.txt
+
+      - name: Build extensions
+        run: |
+          python setup.py build_ext --inplace
+
+      - name: Build wheels
+        run: |
+          python setup.py bdist_wheel
+
+      - name: Repair wheels with repairwheel
+        run: |
+          for whl in dist/*.whl; do
+            repairwheel -o dist/ "$whl"
+            rm "$whl"
+          done
+
+      - name: Upload Linux wheels
+        uses: actions/upload-artifact@v2
+        with:
+          name: wheels-linux
+          path: dist/*.whl
+
+  build_wheels_windows:
+    runs-on: windows-latest
+    strategy:
+      matrix:
+        python-version: [3.7, 3.8, 3.9, 3.10, 3.11, 3.12]
+
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v2
+
+      - name: Set up Python ${{ matrix.python-version }}
+        uses: actions/setup-python@v2
+        with:
+          python-version: ${{ matrix.python-version }}
+
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install -r requirements.txt
+          pip install -r requirements_dev.txt
+          pip install -r requirements_doc.txt
+          pip install -r requirements_test.txt
+
+      - name: Build extensions
+        run: |
+          python setup.py build_ext --inplace
+
+      - name: Build wheels
+        run: |
+          python setup.py bdist_wheel
+
+      - name: Upload Windows wheels
+        uses: actions/upload-artifact@v2
+        with:
+          name: wheels-windows
+          path: dist/*.whl
+
+  build_wheels_macos:
+    runs-on: macos-latest
+    strategy:
+      matrix:
+        python-version: [3.7, 3.8, 3.9, 3.10, 3.11, 3.12]
+
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v2
+
+      - name: Set up Python ${{ matrix.python-version }}
+        uses: actions/setup-python@v2
+        with:
+          python-version: ${{ matrix.python-version }}
+
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install -r requirements.txt
+          pip install -r requirements_dev.txt
+          pip install -r requirements_doc.txt
+          pip install -r requirements_test.txt
+
+      - name: Build extensions
+        run: |
+          python setup.py build_ext --inplace
+
+      - name: Build wheels
+        run: |
+          python setup.py bdist_wheel
+
+      - name: Upload MacOS wheels
+        uses: actions/upload-artifact@v2
+        with:
+          name: wheels-macos
+          path: dist/*.whl
+
+  publish_wheels:
+    needs: [build_wheels_linux, build_wheels_windows, build_wheels_macos]
[…]
