+++ .github/workflows/release.yml
@@ -9,7 +9,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        python-version: ["3.9", "3.10", "3.11"]
+        python-version: ["3.9", "3.10", "3.11", "3.12"]
 
     steps:
       - name: Checkout code
@@ -23,10 +23,10 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install -r requirements.txt
-          pip install -r requirements_dev.txt
-          pip install -r requirements_doc.txt
-          pip install -r requirements_test.txt
+          python -m pip install -r requirements.txt
+          python -m pip install -r requirements_dev.txt
+          python -m pip install -r requirements_doc.txt
+          python -m pip install -r requirements_test.txt
 
       - name: Build extensions
         run: |
@@ -53,7 +53,7 @@ jobs:
     runs-on: windows-latest
     strategy:
       matrix:
-        python-version: ["3.9", "3.10", "3.11"]
+        python-version: ["3.9", "3.10", "3.11", "3.12"]
 
     steps:
       - name: Checkout code
@@ -67,10 +67,10 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install -r requirements.txt
-          pip install -r requirements_dev.txt
-          pip install -r requirements_doc.txt
-          pip install -r requirements_test.txt
+          python -m pip install -r requirements.txt
+          python -m pip install -r requirements_dev.txt
+          python -m pip install -r requirements_doc.txt
+          python -m pip install -r requirements_test.txt
 
       - name: Build extensions
         run: |
@@ -90,7 +90,7 @@ jobs:
     runs-on: macos-latest
     strategy:
       matrix:
-        python-version: ["3.9", "3.10", "3.11"]
+        python-version: ["3.9", "3.10", "3.11", "3.12"]
 
     steps:
       - name: Checkout code
@@ -104,10 +104,10 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install -r requirements.txt
-          pip install -r requirements_dev.txt
-          pip install -r requirements_doc.txt
-          pip install -r requirements_test.txt
+          python -m pip install -r requirements.txt
+          python -m pip install -r requirements_dev.txt
+          python -m pip install -r requirements_doc.txt
+          python -m pip install -r requirements_test.txt
 
       - name: Build extensions
         run: |
