+++ .github/workflows/release.yml
@@ -34,7 +34,7 @@ jobs:
 
       - name: Build wheel
         run: |
-          python -m pip install build
+          python -m pip install build Cython
           python -m build --wheel
 
       - name: Repair wheels with repairwheel
@@ -79,7 +79,7 @@ jobs:
 
       - name: Build wheel
         run: |
-          python -m pip install build
+          python -m pip install build Cython
           python -m build --wheel
 
       - name: Upload Windows wheels
@@ -117,7 +117,7 @@ jobs:
 
       - name: Build wheel
         run: |
-          python -m pip install build
+          python -m pip install build Cython
           python -m build --wheel
 
       - name: Upload MacOS wheels
