+++ .github/workflows/release.yml
@@ -34,7 +34,8 @@ jobs:
 
       - name: Build wheel
         run: |
-          python -m pip install build Cython
+          python -m pip install build
+          python -m pip install Cython
           python -m build --wheel
 
       - name: Repair wheels with repairwheel
@@ -79,7 +80,8 @@ jobs:
 
       - name: Build wheel
         run: |
-          python -m pip install build Cython
+          python -m pip install build 
+          python -m pip install Cython
           python -m build --wheel
 
       - name: Upload Windows wheels
@@ -117,7 +119,8 @@ jobs:
 
       - name: Build wheel
         run: |
-          python -m pip install build Cython
+          python -m pip install build
+          python -m pip install Cython
           python -m build --wheel
 
       - name: Upload MacOS wheels
