+++ .github/workflows/release.yml
@@ -23,6 +23,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
+          python -m pip install setuptools wheel
           python -m pip install -r requirements.txt
           python -m pip install -r requirements_dev.txt
           python -m pip install -r requirements_doc.txt
@@ -69,6 +70,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
+          python -m pip install setuptools wheel
           python -m pip install -r requirements.txt
           python -m pip install -r requirements_dev.txt
           python -m pip install -r requirements_doc.txt
@@ -108,6 +110,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
+          python -m pip install setuptools wheel
           python -m pip install -r requirements.txt
           python -m pip install -r requirements_dev.txt
           python -m pip install -r requirements_doc.txt
