### 2024-07-16 aadya940/chainopy#20 @braniii
And here some additional notes about the comment on the statement of need:

While the package appears to be designed for improved speed compared to pyDTMC, I noticed that generating Markov chain Monte Carlo trajectories with chainopy seems to be relatively slow. In my experience, using alternative tools like deeptime-ml or msmhelper can provide a significant performance boost, sometimes improving speed by 1-3 orders of magnitude, depending on the number of steps required (see attachment).
Is this feature important to your community and if, how many steps are typically needed?

To run this benchmark, you need to run first
``` bash
python -m pip install msmhelper
```

[benchmarks_simulate_mh.zip](https://github.com/user-attachments/files/16247776/benchmarks_simulate_mh.zip)
