### 2024-07-16 aadya940/chainopy#18 @braniii
@aadya940 Sorry for this lengthy review process. But I've finally finished the review. If you have any further questions, feel free to write me.
