### 2024-07-16 aadya940/chainopy#20 @aadya940
Hi @braniii,

First off, I’d like to address the claim that chainopy is designed to be faster than packages that rely solely on `Numpy + Python`, such as `PyDTMC`. This is because Python loops tend to be slow due to its dynamic nature, and Markov chains involve many iterative algorithms (see [this Stack Overflow answer](https://stackoverflow.com/a/8097669)).

I observed that markov sections of deeptime are written in pure C (https://github.com/deeptime-ml/deeptime) which is always faster than Cython. While a library like MSM Helper might benefit from Numba's JIT compilation due to its access to runtime variables, which can offer performance advantages over AOT compilers like Cython, speed is only one aspect of the challenge. Memory management is equally crucial. For instance, a bigram model like Markov Chain can lead to a significant increase in the size of the transition matrix. A corpus with just 100 unique words could consume 78.125MB (considering `double` dtype). Additionally, as the transition matrix becomes more sparse, it’s important to account for sparsity during model saving, which I don’t believe msmhelper does while we do. We also plan to add support for memory-mapped arrays for those that don't fit into memory (see issue #11).

Moreover, I’ve developed chainopy with specific use cases in mind, such as text modeling and stock prediction. While MCMC applications are undoubtedly significant, there are already highly optimized packages like `PyMC`, which are built on tensor libraries and graph compilers like `PyTensor` and `Aesara`. Therefore, a more appropriate comparison would be against `PyMC`.

Thank you for your attention to these points. I hope I've addressed your queries with utmost respected to other libraries and its authors.
