### 2024-07-16 aadya940/chainopy#20 @aadya940
The comparison with pyDTMC pertains strictly to the core aspects of Markov Chains. The primary goal of chainopy is to expand upon these functionalities, enabling the implementation of advanced workflows such as Markov Switching Models, Markov Chain Neural Networks, and Hidden Markov Models (HMMs). These capabilities set chainopy apart from pyDTMC, offering a broader range of applications and increased flexibility for complex modeling tasks.
