+++ .github/workflows/release.yml
@@ -24,10 +24,17 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           python -m pip install setuptools wheel
+          python -m pip install Cython
           python -m pip install -r requirements.txt
           python -m pip install -r requirements_dev.txt
           python -m pip install -r requirements_doc.txt
           python -m pip install -r requirements_test.txt
+          python -m pip install build repairwheel
+
+      - name: Clean dist folder
+        run: |
+          rm -rf dist
+          mkdir dist
 
       - name: Build extensions
         run: |
@@ -35,8 +42,6 @@ jobs:
 
       - name: Build wheel
         run: |
-          python -m pip install build
-          python -m pip install Cython
           python -m build --wheel
 
       - name: Repair wheels with repairwheel
@@ -47,7 +52,7 @@ jobs:
           done
 
       - name: Upload Linux wheels
-        uses: actions/upload-artifact@v2
+        uses: actions/upload-artifact@v3
         with:
           name: wheels-linux
           path: dist/*.whl
@@ -71,10 +76,17 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           python -m pip install setuptools wheel
+          python -m pip install Cython
           python -m pip install -r requirements.txt
           python -m pip install -r requirements_dev.txt
           python -m pip install -r requirements_doc.txt
           python -m pip install -r requirements_test.txt
+          python -m pip install build
+
+      - name: Clean dist folder
+        run: |
+          rm -rf dist
+          mkdir dist
 
       - name: Build extensions
         run: |
@@ -82,12 +94,10 @@ jobs:
 
       - name: Build wheel
         run: |
-          python -m pip install build 
-          python -m pip install Cython
           python -m build --wheel
 
       - name: Upload Windows wheels
-        uses: actions/upload-artifact@v2
+        uses: actions/upload-artifact@v3
         with:
           name: wheels-windows
           path: dist/*.whl
@@ -111,10 +121,17 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           python -m pip install setuptools wheel
+          python -m pip install Cython
           python -m pip install -r requirements.txt
           python -m pip install -r requirements_dev.txt
           python -m pip install -r requirements_doc.txt
           python -m pip install -r requirements_test.txt
+          python -m pip install build
+
+      - name: Clean dist folder
+        run: |
+          rm -rf dist
+          mkdir dist
 
       - name: Build extensions
         run: |
@@ -122,12 +139,10 @@ jobs:
 
       - name: Build wheel
         run: |
-          python -m pip install build
-          python -m pip install Cython
           python -m build --wheel
 
       - name: Upload MacOS wheels
-        uses: actions/upload-artifact@v2
+        uses: actions/upload-artifact@v3
         with:
           name: wheels-macos
           path: dist/*.whl
@@ -137,17 +152,17 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - name: Download Linux wheels
-        uses: actions/download-artifact@v2
+        uses: actions/download-artifact@v3
         with:
           name: wheels-linux
 
       - name: Download Windows wheels
-        uses: actions/download-artifact@v2
+        uses: actions/download-artifact@v3
         with:
           name: wheels-windows
 
       - name: Download MacOS wheels
-        uses: actions/download-artifact@v2
+        uses: actions/download-artifact@v3
         with:
           name: wheels-macos
 
