+++ .github/workflows/release.yml
@@ -22,14 +22,14 @@ jobs:
 
       - name: Install dependencies
         run: |
-          python -m pip install --upgrade pip
-          python -m pip install setuptools wheel
-          python -m pip install Cython
-          python -m pip install -r requirements.txt
-          python -m pip install -r requirements_dev.txt
-          python -m pip install -r requirements_doc.txt
-          python -m pip install -r requirements_test.txt
-          python -m pip install build repairwheel
+          python3 -m pip install --upgrade pip
+          python3 -m pip install setuptools wheel
+          python3 -m pip install Cython
+          python3 -m pip install -r requirements.txt
+          python3 -m pip install -r requirements_dev.txt
+          python3 -m pip install -r requirements_doc.txt
+          python3 -m pip install -r requirements_test.txt
+          python3 -m pip install build repairwheel
 
       - name: Clean dist folder
         run: |
@@ -38,11 +38,11 @@ jobs:
 
       - name: Build extensions
         run: |
-          python setup.py build_ext --inplace
+          python3 setup.py build_ext --inplace
 
       - name: Build wheel
         run: |
-          python -m build --wheel
+          python3 -m build --wheel
 
       - name: Repair wheels with repairwheel
         run: |
@@ -74,14 +74,14 @@ jobs:
 
       - name: Install dependencies
         run: |
-          python -m pip install --upgrade pip
-          python -m pip install setuptools wheel
-          python -m pip install Cython
-          python -m pip install -r requirements.txt
-          python -m pip install -r requirements_dev.txt
-          python -m pip install -r requirements_doc.txt
-          python -m pip install -r requirements_test.txt
-          python -m pip install build
+          python3 -m pip install --upgrade pip
+          python3 -m pip install setuptools wheel
+          python3 -m pip install Cython
+          python3 -m pip install -r requirements.txt
+          python3 -m pip install -r requirements_dev.txt
+          python3 -m pip install -r requirements_doc.txt
+          python3 -m pip install -r requirements_test.txt
+          python3 -m pip install build
 
       - name: Clean dist folder
         run: |
@@ -90,11 +90,11 @@ jobs:
 
       - name: Build extensions
         run: |
-          python setup.py build_ext --inplace
+          python3 setup.py build_ext --inplace
 
       - name: Build wheel
         run: |
-          python -m build --wheel
+          python3 -m build --wheel
 
       - name: Upload Windows wheels
         uses: actions/upload-artifact@v3
@@ -119,14 +119,14 @@ jobs:
 
       - name: Install dependencies
         run: |
-          python -m pip install --upgrade pip
-          python -m pip install setuptools wheel
-          python -m pip install Cython
-          python -m pip install -r requirements.txt
-          python -m pip install -r requirements_dev.txt
-          python -m pip install -r requirements_doc.txt
-          python -m pip install -r requirements_test.txt
-          python -m pip install build
+          python3 -m pip install --upgrade pip
+          python3 -m pip install setuptools wheel
+          python3 -m pip install Cython
+          python3 -m pip install -r requirements.txt
+          python3 -m pip install -r requirements_dev.txt
+          python3 -m pip install -r requirements_doc.txt
+          python3 -m pip install -r requirements_test.txt
+          python3 -m pip install build
 
       - name: Clean dist folder
         run: |
@@ -135,11 +135,11 @@ jobs:
 
       - name: Build extensions
         run: |
-          python setup.py build_ext --inplace
+          python3 setup.py build_ext --inplace
 
       - name: Build wheel
         run: |
-          python -m build --wheel
+          python3 -m build --wheel
 
[…]
