### 2024-07-16 f98a93c
setup.py fixes
