### 2024-07-17 aadya940/chainopy#4 @MichalisPanayides
Yes. That looks a lot better now. Will check off the final item in the checklist now 👍
