+++ .github/workflows/release.yml
@@ -5,166 +5,66 @@ on:
     types: [published]
 
 jobs:
-  build_wheels_linux:
-    runs-on: ubuntu-latest
+  build_wheel_multiplatform:
+    name: Build wheels on ${{ matrix.os }}
+    runs-on: ${{ matrix.os }}
     strategy:
       matrix:
-        python-version: ["3.9", "3.10", "3.11", "3.12"]
+        os: [ubuntu-latest, windows-latest, macos-13, macos-14]
 
     steps:
-      - name: Checkout code
-        uses: actions/checkout@v4
-
-      - name: Set up Python ${{ matrix.python-version }}
-        uses: actions/setup-python@v5
+      - uses: actions/checkout@v4
+      
+      - uses: actions/setup-python@v5
         with:
-          python-version: ${{ matrix.python-version }}
+          python-version: '3.x'
 
       - name: Install dependencies
         run: |
-          python3 -m pip install --upgrade pip
-          python3 -m pip install setuptools wheel
-          python3 -m pip install Cython
-          python3 -m pip install -r requirements.txt
-          python3 -m pip install -r requirements_dev.txt
-          python3 -m pip install -r requirements_doc.txt
-          python3 -m pip install -r requirements_test.txt
-          python3 -m pip install build repairwheel
+          python -m pip install -r requirements.txt
+          python -m pip install -r requirements_dev.txt
 
-      - name: Clean dist folder
-        run: |
-          rm -rf dist
-          mkdir dist
+      - name: Build extensions in place
+        run: python setup.py build_ext --inplace
 
-      - name: Build extensions
-        run: |
-          python3 setup.py build_ext --inplace
-
-      - name: Build wheel
-        run: |
-          python3 -m build --wheel
-
-      - name: Repair wheels with repairwheel
-        run: |
-          for whl in dist/*.whl; do
-            repairwheel -o dist/ "$whl"
-            rm "$whl"
-          done
-
-      - name: Upload Linux wheels
-        uses: actions/upload-artifact@v3
+      - name: Install cibuildwheel
+        run: python -m pip install cibuildwheel==2.19.2
+      
+      - name: Build wheels
+        run: python -m cibuildwheel --output-dir dist
+      
+      - uses: actions/upload-artifact@v4
         with:
-          name: wheels-linux
-          path: dist/*.whl
-
-  build_wheels_windows:
-    runs-on: windows-latest
-    strategy:
-      matrix:
-        python-version: ["3.9", "3.10", "3.11", "3.12"]
-
-    steps:
-      - name: Checkout code
-        uses: actions/checkout@v4
-
-      - name: Set up Python ${{ matrix.python-version }}
-        uses: actions/setup-python@v5
-        with:
-          python-version: ${{ matrix.python-version }}
-
-      - name: Install dependencies
-        run: |
-          python3 -m pip install --upgrade pip
-          python3 -m pip install setuptools wheel
-          python3 -m pip install Cython
-          python3 -m pip install -r requirements.txt
-          python3 -m pip install -r requirements_dev.txt
-          python3 -m pip install -r requirements_doc.txt
-          python3 -m pip install -r requirements_test.txt
-          python3 -m pip install build
-
-      - name: Clean dist folder
-        run: |
-          rm -rf dist
-          mkdir dist
-
-      - name: Build extensions
-        run: |
-          python3 setup.py build_ext --inplace
-
-      - name: Build wheel
-        run: |
-          python3 -m build --wheel
-
-      - name: Upload Windows wheels
-        uses: actions/upload-artifact@v3
-        with:
-          name: wheels-windows
-          path: dist/*.whl
-
-  build_wheels_macos:
-    runs-on: macos-latest
-    strategy:
-      matrix:
[…]
