+++ .github/workflows/release.yml
@@ -25,7 +25,9 @@ jobs:
           python -m pip install -r requirements_dev.txt
 
       - name: Build extensions in place
-        run: python setup.py build_ext --inplace
+        run: |
+          python -m pip install setuptools wheel
+          python setup.py build_ext --inplace
 
       - name: Install cibuildwheel
         run: python -m pip install cibuildwheel==2.19.2
