+++ .github/workflows/release.yml
@@ -26,15 +26,17 @@ jobs:
 
       - name: Build extensions in place
         run: |
-          python -m pip install setuptools wheel
+          python -m pip install setuptools wheel 
           python setup.py build_ext --inplace
 
       - name: Install cibuildwheel
         run: python -m pip install cibuildwheel==2.19.2
       
       - name: Build wheels
-        run: python -m cibuildwheel --output-dir dist
-      
+        run: |
+          python -m cibuildwheel --output-dir dist
+          python -m pip install Cython
+
       - uses: actions/upload-artifact@v4
         with:
           name: wheels-${{ matrix.os }}
