+++ .github/workflows/release.yml
@@ -10,6 +10,7 @@ jobs:
     runs-on: ${{ matrix.os }}
     strategy:
       matrix:
+        python: ["cp39", "cp310", "cp311", "cp312"]
         os: [ubuntu-latest, windows-latest, macos-13, macos-14]
 
     steps:
@@ -35,7 +36,8 @@ jobs:
       - name: Build wheels
         run: |
           python -m cibuildwheel --output-dir dist
-          python -m pip install Cython
+        env:
+          CIBW_BUILD: "${{ matrix.python }}-*"
 
       - uses: actions/upload-artifact@v4
         with:
