+++ .github/workflows/release.yml
@@ -37,6 +37,7 @@ jobs:
         run: |
           python -m cibuildwheel --output-dir dist
         env:
+          CIBW_BEFORE_BUILD: python -m pip install -r requirements.txt 
           CIBW_BUILD: "${{ matrix.python }}-*"
 
       - uses: actions/upload-artifact@v4
