+++ .github/workflows/release.yml
@@ -37,7 +37,9 @@ jobs:
         run: |
           python -m cibuildwheel --output-dir dist
         env:
-          CIBW_BEFORE_BUILD: pip install -r requirements.txt 
+          CIBW_BEFORE_BUILD: >
+            python -m pip install --upgrade pip
+            pip install -r requirements.txt 
           CIBW_BUILD: "${{ matrix.python }}-*"
 
       - uses: actions/upload-artifact@v4
