+++ .github/workflows/release.yml
@@ -39,7 +39,8 @@ jobs:
           python -m cibuildwheel --output-dir dist
         env:
           CIBW_BEFORE_BUILD: >
-            python -m pip install --upgrade pip && pip install -r requirements.txt 
+            python -m pip install --upgrade pip && 
+            pip install -r requirements.txt 
           CIBW_BUILD: "${{ matrix.python }}-*"
 
       - uses: actions/upload-artifact@v4
