+++ .github/workflows/release.yml
@@ -38,8 +38,8 @@ jobs:
         run: |
           python -m cibuildwheel --output-dir dist
         env:
-          CIBW_BEFORE_BUILD: >
-            python -m pip install --upgrade pip && 
+          CIBW_BEFORE_BUILD: |
+            python -m pip install --upgrade pip 
             pip install -r requirements.txt 
           CIBW_BUILD: "${{ matrix.python }}-*"
 
