+++ .github/workflows/release.yml
@@ -40,7 +40,7 @@ jobs:
         env:
           CIBW_BEFORE_BUILD: |
             python -m pip install --upgrade pip 
-            pip install -r requirements.txt 
+            python -m pip install -r requirements.txt 
           CIBW_BUILD: "${{ matrix.python }}-*"
 
       - uses: actions/upload-artifact@v4
