+++ .github/workflows/release.yml
@@ -52,6 +52,12 @@ jobs:
         with:
           python-version: "3.9"
 
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install numpy
+          pip install cibuildwheel
+          
       - name: Install twine
         run: |
           python -m pip install --upgrade pip
