+++ .github/workflows/release.yml
@@ -29,7 +29,7 @@ jobs:
       - name: Build wheels
         run: cibuildwheel --output-dir wheelhouse
         env:
-          CIBW_BUILD: cp3?-*
+          CIBW_BUILD: "cp3{9,10,11,12}-*"
           CIBW_SKIP: "pp*"
 
       - name: Upload wheels to GitHub
@@ -57,7 +57,7 @@ jobs:
           python -m pip install --upgrade pip
           pip install numpy
           pip install cibuildwheel
-          
+
       - name: Install twine
         run: |
           python -m pip install --upgrade pip
