### 2024-07-19 aadya940/chainopy#17 @aadya940
Hi @braniii , the package is downloadable via `pip install git+`. I've also written a `release.yml` but it has some issues at the moment. :))
