+++ JOSS/paper.md
@@ -14,6 +14,12 @@ affiliations:
 bibliography: paper.bib
 ---
 
+### Summmary
+One of the easiest and most intuitive way to model time series data like Stock Prices, Text Words etc is a Markov
+Chain. ChainoPy makes it easy for you to model time series data using models like Markov Chains, Markov Switching 
+Models etc while taking computational aspects like speed and memory into account. Apart from that, you could also
+model behavior of probablistic models like Markov Chains through Neural Networks which are widely considered 
+deterministic using the `MarkovChainNeuralNetwork` class.
 
 ### Statement of Need
 There are significant limitations in current Markov Chain packages that rely solely on NumPy [@harris2020array] and Python for implementation. Markov Chains often require iterative convergence-based algorithms [@rosenthal1995convergence], where Python's dynamic typing, Global Interpreter Lock (GIL), and garbage collection can hinder potential performance improvements like parallelism. To address these issues, we enhance our library with extensions like Cython and Numba for efficient algorithm implementation. Additionally, we introduce a Markov Chain Neural Network [@awiszus2018markov] that simulates given Markov Chains while preserving statistical properties from the training data. This approach eliminates the need for post-processing steps such as sampling from the outcome distribution while giving neural networks stochastic properties rather than deterministic behavior. Finally, we implement the famous Markov Switching Models [@hamilton2010regime] which are one of the fundamental and widely used models in applications such as Stock Market price prediction.
