+++ setup.py
@@ -61,11 +61,5 @@ setup(
         "numba",
         "statsmodels",
     ],
-    classifiers=[
-        "Programming Language :: Python",
-        "Programming Language :: Python :: 3",
-        "License :: OSI Approved :: MIT License",
-        "Operating System :: OS Independent",
-    ],
     python_requires=">=3.9",
 )
