### 2024-07-20 2e19769
setup.py fixes
