### 2024-07-20 87bcde7
readthedocs yml fixes
