+++ .readthedocs.yaml
@@ -6,8 +6,6 @@ build:
     python: "3.10"
   jobs:
     pre_build:
-      - sudo apt-get install python3.10-dev
-      - sudo apt install build-essential
       - python -m pip install build
       - python -m build
       - sphinx-apidoc --private -o docs/source chainopy/
