### 2024-07-20 983e6de
expose _markov_chain, _nn, _markov_switching through the public API
