+++ chainopy/__init__.py
@@ -1,6 +1,6 @@
-from ._markov_chain import MarkovChain
-from ._nn import MarkovChainNeuralNetwork, divergance_analysis
-from ._markov_switching import MarkovSwitchingModel
+from .markov_chain import MarkovChain
+from .nn import MarkovChainNeuralNetwork, divergance_analysis
+from .markov_switching import MarkovSwitchingModel
 
 __all__ = [
     "MarkovChain",
