+++ .readthedocs.yaml
@@ -6,8 +6,9 @@ build:
     python: "3.10"
   jobs:
     pre_build:
-      - python -m pip install build
-      - python -m build
+      - python -m pip install -r requirements.txt
+      - python -m pip install -r requirements_doc.txt
+      - python setup.py build_ext --inplace
       - sphinx-apidoc --private -o docs/source chainopy/
 
 python:
