+++ docs/source/conf.py
@@ -60,6 +60,7 @@ autodoc_default_options = {
     "members": True,
     "undoc-members": True,
     "special-members": "__init__",
+    "private-members": False,
 }
 
 autosummary_generate = True
