+++ docs/source/conf.py
@@ -51,6 +51,10 @@ exclude_patterns = [
     ".DS_Store",
     "chainopy.tests.rst",
     "chainopy._backend.rst",
+    "chainopy._caching.rst",
+    "chainopy._fileio.rst",
+    "chainopy._visualization.rst",
+    "chainopy._exceptions.rst",
     "modules.rst",
     "setup.rst",
     "_templates",
