+++ chainopy/markov_chain.py
@@ -4,10 +4,10 @@ from typing import List, Union
 import numpy as np
 import numba
 
-from ._exceptions import handle_exceptions
+from ._exceptions import _handle_exceptions
 from ._visualizations import _visualize_tpm, _visualize_chain
-from ._fileio import _save_model_markovchain, _load_model_markovchain, load_text
-from ._caching import cache
+from ._fileio import _save_model_markovchain, _load_model_markovchain, _load_text
+from ._caching import _cache
 from ._backend import (
     _simulate,
     _absorbing,
@@ -96,8 +96,8 @@ class MarkovChain:
                                             (TPM) {p} must sum to 1."
                     )
 
-    @handle_exceptions
-    @cache(class_method=True)
+    @_handle_exceptions
+    @_cache(class_method=True)
     def fit(self, data: Union[str, list], epsilon: float = 1e-16) -> np.ndarray:
         """
         Learn Transition Matrix from Sequence (list or str) of Data.
@@ -130,7 +130,7 @@ class MarkovChain:
         self._validate_transition_matrix(self.tpm, self.states, self.epsilon)
         return _tpm
 
-    @handle_exceptions
+    @_handle_exceptions
     def simulate(self, initial_state: str, n_steps: int) -> List[str]:
         """
         Simulate the Markov Chain for `n_steps` steps.
@@ -149,7 +149,7 @@ class MarkovChain:
         return _simulate._simulate_cython(self.states, self.tpm, initial_state, n_steps)
 
     @staticmethod
-    @numba.jit(nopython=True, cache=True)
+    @numba.jit(nopython=True, _cache=True)
     def _vectorize(states: List[str], initial_state: str) -> np.ndarray:
         if initial_state in states:
             init = states.index(initial_state)
@@ -183,7 +183,7 @@ class MarkovChain:
         initial_vect = self._vectorize(self.states, initial_state)
         return self.states[np.argmax(initial_vect @ self.tpm)]
 
-    @handle_exceptions
+    @_handle_exceptions
     def nstep_distribution(self, n_steps: int) -> np.ndarray:
         """
         Calculates the distribution of the Markov Chain after n-steps.
@@ -211,7 +211,7 @@ class MarkovChain:
 
         return result if is_eigendecom else np.linalg.matrix_power(self.tpm, n_steps)
 
-    @cache(class_method=True)
+    @_cache(class_method=True)
     def is_ergodic(self) -> bool:
         """
         Checks if the Markov chain is ergodic.
@@ -222,7 +222,7 @@ class MarkovChain:
         """
         return self.is_irreducible() and self.is_aperiodic()
 
-    @cache(class_method=True)
+    @_cache(class_method=True)
     def is_symmetric(self) -> bool:
         """
         Checks if the Markov chain is symmetric.
@@ -233,7 +233,7 @@ class MarkovChain:
         """
         return np.allclose(self.tpm, self.tpm.transpose())
 
-    @cache(class_method=True)
+    @_cache(class_method=True)
     def stationary_dist(self) -> np.ndarray:
         """
         Returns the stationary distribution of the Markov chain.
@@ -245,8 +245,8 @@ class MarkovChain:
         tpm_T = self.tpm.transpose()
         return _stationary_dist.cython_stationary_dist(tpm_T)
 
-    @handle_exceptions
-    @cache(class_method=True)
+    @_handle_exceptions
+    @_cache(class_method=True)
     def is_communicating(self, state1: str, state2: str, threshold: int = 1000) -> bool:
         """
         Checks if two states are communicating.
@@ -268,7 +268,7 @@ class MarkovChain:
             self.tpm, self.states, state1, state2, threshold
         )
 
-    @cache(class_method=True)
+    @_cache(class_method=True)
[…]
