+++ docs/source/conf.py
@@ -55,6 +55,12 @@ exclude_patterns = [
     "chainopy._fileio.rst",
     "chainopy._visualization.rst",
     "chainopy._exceptions.rst",
+    "chainopy.chainopy.tests.rst",
+    "chainopy.chainopy._backend.rst",
+    "chainopy.chainopy._caching.rst",
+    "chainopy.chainopy._fileio.rst",
+    "chainopy.chainopy._visualization.rst",
+    "chainopy.chainopy._exceptions.rst",
     "modules.rst",
     "setup.rst",
     "_templates",
