+++ README.md
@@ -14,6 +14,12 @@ A Python 🐍 Package for Markov Chains, Markov Chain Neural Networks and Markov
 
 # How to Install ChainoPy?
 
+### Using pip
+
+```bash
+pip install git+https://github.com/aadya940/chainopy
+```
+
 ### Build from Source
 
 Before you begin, ensure you have the following installed on your system:
