### 2024-07-23 c53d27f
Delete .github/workflows/release.yml
