+++ .github/workflows/release.yml
@@ -1,49 +1,0 @@
-name: Build and Repair Wheels on Release
-
-on:
-  release:
-    types: [published]
-
-jobs:
-  build:
-    runs-on: ${{ matrix.os }}
-    strategy:
-      matrix:
-        os: [ubuntu-latest, macos-latest, windows-latest]
-        python-version: ["3.7", "3.8", "3.9", "3.10", "3.11", "3.12"]
-
-    steps:
-    - uses: actions/checkout@v4
-
-    - name: Set up Python ${{ matrix.python-version }}
-      uses: actions/setup-python@v4
-      with:
-        python-version: ${{ matrix.python-version }}
-
-    - name: Install build dependencies
-      run: |
-        python -m pip install --upgrade pip
-        pip install build numpy Cython scipy
-
-    - name: Build wheels
-      run: python -m build --wheel --outdir dist
-
-    - name: Install repairwheel (Linux only)
-      if: runner.os == 'Linux'
-      run: pip install repairwheel
-
-    - name: Repair wheels (Linux only)
-      if: runner.os == 'Linux'
-      run: |
-        for whl in dist/*.whl; do
-          repairwheel -o dist "$whl"
-          rm "$whl"
-        done
-
-    - name: Upload wheels to PyPI
-      env:
-        TWINE_USERNAME: ${{ secrets.PYPI_USERNAME }}
-        TWINE_PASSWORD: ${{ secrets.PYPI_PASSWORD }}
-      run: |
-        pip install twine
-        twine upload dist/*.whl
