+++ chainopy/markov_chain.py
@@ -16,9 +16,6 @@ from ._backend import (
     _is_communicating,
 )
 
-__all__ = ["MarkovChain"]
-
-
 class MarkovChain:
     """
     A class containing Fundamental Functions for Discrete Time Markov Chains.
