### 2024-07-28 a8f16d9
Update markov_switching.py
