### 2024-07-29 aadya940/chainopy#17 @aadya940
Hey @braniii , Is chainopy installable for you? `pip install chainopy`
