+++ JOSS/paper.bib
@@ -12,6 +12,15 @@
   howpublished = {\url{https://github.com/TommasoBelluzzo/PyDTMC}}
   }
 
+@misc{simple-markov,
+  title = {PyDTMC},
+  howpublished = {\url{https://github.com/Mandragorian/simple-markov}}
+  }
+
+@misc{mchmm,
+  title = {PyDTMC},
+  howpublished = {\url{https://github.com/maximtrp/mchmm}}
+  }
 
 @article{behnel2010cython,
   title={Cython: The best of both worlds},
