+++ JOSS/paper.bib
@@ -13,12 +13,12 @@
   }
 
 @misc{simple-markov,
-  title = {PyDTMC},
+  title = {simple-markov},
   howpublished = {\url{https://github.com/Mandragorian/simple-markov}}
   }
 
 @misc{mchmm,
-  title = {PyDTMC},
+  title = {mchmm},
   howpublished = {\url{https://github.com/maximtrp/mchmm}}
   }
 
