### 2024-07-29 aadya940/chainopy#20 @aadya940
@braniii I've mentioned a few packages in the text modelling domain that use pure python and numpy.
