### 2024-07-30 aadya940/chainopy#17 @braniii
> Hey @braniii , Is chainopy installable for you? `pip install chainopy`

You fixed it for py3.12.  But now there are no wheels for py3.11, py3.10 etc. So it works for py3.12 only.
