### 2024-07-30 aadya940/chainopy#18 @braniii
> The review has been really helpful to target sections which were lacking before. Thanks for the comprehensive review :))

You are welcome :)
