### 2024-07-30 aadya940/chainopy#20 @braniii
> @braniii Also I added a few lines in the Statement of Need. I hope everything looks good to you now :))

These are really helping to underline the statement of chainopy :+1:
