### 2024-07-31 4203740
Add class level docstring
