### 2024-07-31 aadya940/chainopy#17 @aadya940
@braniii I was using Numba earlier but I removed it as a part of a recent cleanup PR since the function it was used in wasn't contributing much to performance. Removed the word `Numba` from the paper also :))
