### 2024-07-31 aadya940/chainopy#18 @aadya940
@braniii
For point three:
```
 Add examples for the advertised functionality in the docs.
```
I believe I've added plenty examples in https://github.com/aadya940/chainopy/blob/master/examples/demo.ipynb ,
So I've just referenced this file at the top of my sphinx documentation and hence am checking the box. Hope it is fine for you :))
