+++ pyproject.toml
@@ -2,7 +2,7 @@
 requires = [
     "setuptools >= 61.0",
     "wheel",
-    "numpy >= 1.26.1",
+    "numpy == 2.0.0",
     "scipy",
     "Cython",
 ]
@@ -21,7 +21,7 @@ authors = [
 keywords = ["markov chain", "stochastic analysis", "python"]
 dependencies = [
     "Cython",
-    "numpy",
+    "numpy == 2.0.0",    
     "scipy",
     "xarray",
     "matplotlib",
