### 2024-07-31 9cf513a
update scipy version
