+++ pyproject.toml
@@ -3,7 +3,7 @@ requires = [
     "setuptools >= 61.0",
     "wheel",
     "numpy == 2.0.0",
-    "scipy",
+    "scipy >= 1.13.0",
     "Cython",
 ]
 build-backend = "setuptools.build_meta"
@@ -22,7 +22,7 @@ keywords = ["markov chain", "stochastic analysis", "python"]
 dependencies = [
     "Cython",
     "numpy == 2.0.0",    
-    "scipy",
+    "scipy >= 1.13.0",
     "xarray",
     "matplotlib",
     "networkx",
