+++ .github/workflows/release.yml
@@ -20,9 +20,24 @@ jobs:
           package-dir: .
           output-dir: dist
           config-file: "./pyproject.toml"
+        env:
+          CIBW_ENVIRONMENT: sudo apt-get install python3-scipy
 
       - uses: actions/upload-artifact@v4
         with:
           name: cibw-wheels-${{ matrix.os }}-${{ strategy.job-index }}
           path: ./dist/*.whl
           retention-days: 5
+
+  download:
+    runs-on: ubuntu-latest
+    needs: build
+
+    steps:
+    - name: Download build artifacts
+      uses: actions/download-artifact@v3
+      with:
+        path: dist/
+
+    - name: List downloaded files
+      run: ls -l dist/
