+++ .github/workflows/release.yml
@@ -20,8 +20,8 @@ jobs:
           package-dir: .
           output-dir: dist
           config-file: "./pyproject.toml"
-        env:
-          CIBW_ENVIRONMENT: sudo apt-get install python3-scipy
+          env:
+            CIBW_ENVIRONMENT: sudo apt-get install python3-scipy
 
       - uses: actions/upload-artifact@v4
         with:
