+++ .github/workflows/release.yml
@@ -21,7 +21,7 @@ jobs:
           output-dir: dist
           config-file: "./pyproject.toml"
           env:
-            CIBW_ENVIRONMENT: sudo apt-get install python3-scipy
+            CIBW_BEFORE_ALL: sudo apt-get install python3-scipy
 
       - uses: actions/upload-artifact@v4
         with:
