+++ .github/workflows/release.yml
@@ -20,7 +20,7 @@ jobs:
           package-dir: .
           output-dir: dist
           config-file: "./pyproject.toml"
-          env:
+        env:
             CIBW_BEFORE_ALL: sudo apt-get install python3-scipy
 
       - uses: actions/upload-artifact@v4
