+++ pyproject.toml
@@ -3,7 +3,6 @@ requires = [
     "setuptools >= 61.0",
     "wheel",
     "numpy == 2.0.0",
-    "scipy >= 1.13.0",
     "Cython",
 ]
 build-backend = "setuptools.build_meta"
