+++ pyproject.toml
@@ -2,7 +2,7 @@
 requires = [
     "setuptools >= 61.0",
     "wheel",
-    "scipy == 1.14.0",
+    "scipy >= 1.13.1",
     "numpy == 2.0.0",
     "Cython",
 ]
