+++ pyproject.toml
@@ -40,4 +40,7 @@ repository = "https://github.com/aadya940/chainopy"
 skip = "pp*"
 
 [tool.cibuildwheel.linux]
-before-all="apt build-dep scipy"
+before-all=[
+    "yum install epel-release",
+    "yum install python3-devel numpy-f2py gcc-gfortran openblas-devel lapack-devel",
+]
