+++ .github/workflows/release.yml
@@ -9,7 +9,7 @@ jobs:
     strategy:
       matrix:
         # macos-13 is an intel runner, macos-14 is apple silicon
-        os: [ubuntu-latest, macos-13, macos-14]
+        os: [winows-latest, ubuntu-latest, macos-13, macos-14]
 
     steps:
       - uses: actions/checkout@v4
