+++ pyproject.toml
@@ -37,7 +37,7 @@ repository = "https://github.com/aadya940/chainopy"
 
 [tool.cibuildwheel]
 # Disable building PyPy wheels on all platforms
-skip = "pp*"
+skip = "pp* *musllinux*"
 
 [tool.cibuildwheel.linux]
 before-all=[
