+++ .github/workflows/release.yml
@@ -28,14 +28,15 @@ jobs:
           retention-days: 5
 
   download:
-    runs-on: ubuntu-latest
-    needs: build_wheels
+      runs-on: ubuntu-latest
+      needs: build_wheels
 
-    steps:
-    - name: Download build artifacts
-      uses: actions/download-artifact@v3
-      with:
-        path: dist/
+      steps:
+        - name: Download build artifacts
+          uses: actions/download-artifact@v3
+          with:
+            name: cibw-wheels-${{ matrix.os }}-${{ strategy.job-index }}
+            path: dist/
 
-    - name: List downloaded files
-      run: ls -l dist/
+        - name: List downloaded files
+          run: ls -l dist/
