### 2024-08-01 aadya940/chainopy#17 @aadya940
@braniii Finally I've uploaded all 16 wheels, check it out here [](https://pypi.org/manage/project/chainopy/releases/)
