### 2024-08-05 aadya940/chainopy#17 @braniii
> Now I managed to pip install but I get following error:
>
> ```
> # setting up a conda env
> $ conda create -n chainopy python
> $ conda activate chainopy
>
> # install chainopy
> $ python -m pip install chainopy
>
> $ python
> Python 3.12.4 | packaged by Anaconda, Inc. | (main, Jun 18 2024, 15:12:24) [GCC 11.2.0] on linux
> Type "help", "copyright", "credits" or "license" for more information.
> >>> import chainopy
>
> A module that was compiled using NumPy 1.x cannot be run in
> NumPy 2.0.1 as it may crash. To support both 1.x and 2.x
> versions of NumPy, modules must be compiled with NumPy 2.0.
> Some module may need to rebuild instead e.g. with 'pybind11>=2.12'.
>
> If you are a user of the module, the easiest solution will be to
> downgrade to 'numpy<2' or try to upgrade the affected module.
> We expect that some modules will need time to support NumPy 2.
>
> Traceback (most recent call last):  File "<stdin>", line 1, in <module>
>   File "/home/.../chainopy/chainopy/__init__.py", line 1, in <module>
>     from .markov_chain import MarkovChain
>   File "/home/.../chainopy/chainopy/markov_chain.py", line 11, in <module>
>     from ._backend import (
>   File "/home/.../chainopy/chainopy/_backend/__init__.py", line 1, in <module>
>     from . import _simulate
> Traceback (most recent call last):
>   File "<stdin>", line 1, in <module>
>   File "/home/.../chainopy/chainopy/__init__.py", line 1, in <module>
>     from .markov_chain import MarkovChain
>   File "/home/.../chainopy/chainopy/markov_chain.py", line 11, in <module>
>     from ._backend import (
>   File "/home/.../chainopy/chainopy/_backend/__init__.py", line 1, in <module>
>     from . import _simulate
>   File "chainopy/_backend/_simulate.pyx", line 1, in init chainopy._backend._simulate
>     import numpy as np
> ImportError: numpy.core.multiarray failed to import (auto-generated because you didn't call 'numpy.import_array()' after cimporting numpy; use '<void>numpy._import_array' to disable if you are certain you don't need it).
> >>>
> ```
>
> PS: I get the same error using `pip install git+...`

For me this error persists and I cannot import/use chainopy.
