### 2024-08-05 aadya940/chainopy#18 @braniii
> @braniii For point three:
>
> ```
>  Add examples for the advertised functionality in the docs.
> ```
>
> I believe I've added plenty examples in https://github.com/aadya940/chainopy/blob/master/examples/demo.ipynb , So I've just referenced this file at the top of my sphinx documentation and hence am checking the box. Hope it is fine for you :))

The issue with githubs ipynb rendering engine is, that it is still not supported in the official app. Could you therefore simply include the notebook as a new page into your docs?
