+++ requirements.txt
@@ -1,5 +1,5 @@
 cython
-numpy==1.26.1
+numpy==2.0.0
 scipy
 xarray
 matplotlib
