### 2024-08-05 1732e9e
Ensure compatability with numpy 2.0
