+++ pyproject.toml
@@ -1,9 +1,9 @@
 [build-system]
 requires = [
+    "numpy == 2.0.0",
     "setuptools >= 61.0",
     "wheel",
     "scipy >= 1.13.1",
-    "numpy == 2.0.0",
     "Cython",
 ]
 build-backend = "setuptools.build_meta"
@@ -20,8 +20,8 @@ authors = [
 ]
 keywords = ["markov chain", "stochastic analysis", "python"]
 dependencies = [
+    "numpy == 2.0.0",
     "Cython",
-    "numpy == 2.0.0",    
     "scipy >= 1.13.0",
     "xarray",
     "matplotlib",
