### 2024-08-05 8ed5f0a
Ensure numpy 2.0 compatibility
