+++ pyproject.toml
@@ -10,7 +10,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "chainopy"
-version = "1.0.1"
+version = "1.0.2"
 description = "A Python Library for Markov Chain based Stochastic Analysis!"
 readme = "README.md"
 requires-python = ">=3.9"
