### 2024-08-05 4178f68
Ensure compatibility with numpy 2.0
