### 2024-08-05 b332c35
Ensure compatibility with numpy 2.0
