### 2024-08-05 7a5293c
Remove support for python 3.9, Add py 3.12 to CI
