### 2024-08-05 72b3bc3
Embed Jupyter Notebook in the docs
