+++ requirements_doc.txt
@@ -3,3 +3,4 @@ sphinx-rtd-theme
 furo
 myst-parser
 sphinx-autodoc-typehints
+nbsphinx
