+++ docs/source/index.rst
@@ -9,13 +9,8 @@ Welcome to Chainopy's documentation!
 Examples
 --------
 
-You can see example usage of chainopy here: 
-
-.. toctree::
-   :maxdepth: 2
-   :caption: Examples:
-
-   examples/demo.ipynb
+You can see example usage of chainopy here: `example notebook`_.
+   .. _example notebook ../../examples/demo.ipynb
 
 Autosummary
 -----------
