### 2024-08-05 aadya940/chainopy#17 @aadya940
@braniii
- Hey, sorry we had an incompatibility with Numpy 2.0.0 which is fixed now. Can you retry?
- Also, pydtmc might not work with numpy 2.0.0 (Incase you're going to run the notebook again locally).
- We also removed support for python 3.9 since chainopy requires networkx and only networkx version 3.3 supports numpy 2.0.0 but networkx doesn't support python 3.9.
- Please let me know if you have any issues, Thanks for your patience :))
