### 2024-08-05 d348e20
Embed IPython notebook in docs
