### 2024-08-05 42c333a
update doc ipython path
