### 2024-08-05 3d92f07
enable myst extensions
