+++ docs/source/conf.py
@@ -80,9 +80,7 @@ myst_enable_extensions = [
     "fieldlist",
     "html_admonition",
     "html_image",
-    "linkify",
     "replacements",
-    "smartquotes",
     "strikethrough",
     "substitution",
     "tasklist",
