+++ README.md
@@ -60,7 +60,7 @@ Create Markov Chains and Markov Chain Neural Networks as follows:
 >>> neural_network = chainopy.MarkovChainNeuralNetwork(mc, num_layers = 5)    # Creates a 5-layered Neural Network that simulates `mc`. 
 ```
 
-![image](https://github.com/aadya940/chainopy/blob/master/figs/Simulation-MCNN-2x2.png)
+![image](https://github.com/user-attachments/assets/ff8801eb-bf0d-44d5-8ff5-706698164634)
 
 Create a Markov Switching Model as follows:
 
