### 2024-08-06 4b25688
add ipython doc
