+++ docs/source/conf.py
@@ -72,20 +72,6 @@ exclude_patterns = [
     "Subpackages.rst",
 ]
 
-myst_enable_extensions = [
-    "amsmath",
-    "colon_fence",
-    "deflist",
-    "dollarmath",
-    "fieldlist",
-    "html_admonition",
-    "html_image",
-    "replacements",
-    "strikethrough",
-    "substitution",
-    "tasklist",
-]
-
 autodoc_default_options = {
     "members": True,
     "undoc-members": False,
