+++ docs/source/index.rst
@@ -11,8 +11,9 @@ Examples
 
 .. toctree::
    :maxdepth: 2
+   :caption: Notebooks
 
-   examples
+   examples/demo.ipynb
 
 Autosummary
 -----------
