+++ docs/source/examples.rst
@@ -1,1 +1,0 @@
-.. include::  ../../examples/demo.ipynb
