### 2024-08-06 aadya940/chainopy#17 @braniii
> @braniii
>
>     * Hey, sorry we had an incompatibility with Numpy 2.0.0 which is fixed now. Can you retry?
>
>     * Also, pydtmc might not work with numpy 2.0.0 (Incase you're going to run the notebook again locally).
>
>     * We also removed support for python 3.9 since chainopy requires networkx and only networkx version 3.3 supports numpy 2.0.0 but networkx doesn't support python 3.9.
>
>     * Please do the installation in a `venv`.
>
>     * Please let me know if you have any issues, Thanks for your patience :))

The current version is still not working. Now with a different error. Please check if the issue is fixed, **before** pinging me.
