### 2024-08-06 #6880 @braniii
@mooniean There are still two open issues (make it installable via pip/pypi and add examples to the docs). Once these are fixed, I will check off the remaining boxes in my review.
