### 2024-08-07 51015c5
jupyter notebook in docs
