+++ docs/source/examples.rst
@@ -1,0 +1,2 @@
+.. include:: demo.ipynb
+   :parser: myst_nb.docutils_
