+++ docs/source/examples.rst
@@ -1,2 +1,1 @@
-.. include:: demo.ipynb
-   :parser: myst_nb.docutils_
+../../examples
