### 2024-08-07 94e7fde
Add pandoc dependency for docs
