+++ requirements_doc.txt
@@ -4,3 +4,4 @@ furo
 myst-parser
 sphinx-autodoc-typehints
 nbsphinx
+pandoc
