### 2024-08-07 bdeb0fa
install pandoc on CI systems
