+++ .readthedocs.yaml
@@ -8,6 +8,8 @@ build:
     pre_build:
       - python -m pip install -r requirements.txt
       - python -m pip install -r requirements_doc.txt
+      - sudo apt install pandoc
+      - python -m pip install pandoc
       - python setup.py build_ext --inplace
       - sphinx-apidoc --private -o docs/source chainopy/
 
