+++ .readthedocs.yaml
@@ -8,7 +8,7 @@ build:
     pre_build:
       - python -m pip install -r requirements.txt
       - python -m pip install -r requirements_doc.txt
-      - sudo apt install pandoc
+      - apt install -y pandoc
       - python -m pip install pandoc
       - python setup.py build_ext --inplace
       - sphinx-apidoc --private -o docs/source chainopy/
