+++ .readthedocs.yaml
@@ -8,9 +8,7 @@ build:
     pre_build:
       - python -m pip install -r requirements.txt
       - python -m pip install -r requirements_doc.txt
-      - apt update -y
-      - apt install -y pandoc
-      - python -m pip install pandoc
+      - python -m pip install pypandoc_binary
       - python setup.py build_ext --inplace
       - sphinx-apidoc --private -o docs/source chainopy/
 
