### 2024-08-07 e0868c1
ipython doc
