### 2024-08-07 aadya940/chainopy#18 @aadya940
@braniii Hello,
- The Jupyter Example Notebooks are now visible inside the docs under Examples section, you could check it out.
- I believe I've also fixed the rendering issue inside the README part.
- Attributes like epsilon being documented twice seems like a Sphinx issue since it doesn't behave well with `__slots__`, check it out here :
[https://stackoverflow.com/questions/78789259/how-to-prevent-sphinx-from-displaying-duplicate-attributes-in-class-documentatio](https://stackoverflow.com/questions/78789259/how-to-prevent-sphinx-from-displaying-duplicate-attributes-in-class-documentatio)
Could we skip it for now till I come up with a good solution?

I hope everything seems good now :)) , Sorry for the delay.
