### 2024-08-09 #6880 @MichalisPanayides
> Hello everyone! I've checked that both reviewers have checked all the boxes. @braniii @MichalisPanayides are you happy to recommend the paper for publication in JOSS?

Yes. I am happy with the changes
