+++ README.md
@@ -9,6 +9,7 @@ A Python 🐍 Package for Markov Chains, Markov Chain Neural Networks and Markov
 - Faster than other libraries (eg: 5x Faster than PyDTMC)
 - First Package to contain functions to build equivalent [Markov Chain Neural Networks](https://openaccess.thecvf.com/content_cvpr_2018_workshops/papers/w42/Awiszus_Markov_Chain_Neural_CVPR_2018_paper.pdf) from Markov Chains.
 - Contains Markov Switching Models for Univariate Time Series Analysis
+- Supports Numpy 2.0.0
   
 
 
