### 2024-08-12 #6880 @mooniean
Hi @aadya940 ! Before we do the post-review checklist, there's only a couple of small corrections for the paper (for consistency purposes in the most part):

There are some capitalisation inconsistencies:
Line 29: writing pyDTMC while previous was PyDTMC (and afterwards too)
Line 36: same with cython and python (Python has to be capitalised)
Line 39: markov chains > Markov chains

And a small typo on line 115: should be packages ("we use the Pytest and PyDTMC package_s_)
