### 2024-08-12 8fe5f45
Fix typing inconsistencies
