### 2024-08-12 63a2496
Fix typing inconsistencies
