### 2024-08-12 #6880 @aadya940
@mooniean I already have a release made:

https://github.com/aadya940/chainopy/releases/tag/v1.0.2
