### 2024-08-12 27db2ac
add doi's and urls
