+++ JOSS/paper.bib
@@ -3,23 +3,27 @@
   author={Awiszus, Maren and Rosenhahn, Bodo},
   booktitle={Proceedings of the IEEE Conference on Computer Vision and Pattern Recognition Workshops},
   pages={2180--2187},
-  year={2018}
+  year={2018},
+  url={https://openaccess.thecvf.com/content_cvpr_2018_workshops/papers/w42/Awiszus_Markov_Chain_Neural_CVPR_2018_paper.pdf}
 }
 
 
 @misc{pydtmc,
   title = {PyDTMC},
-  howpublished = {\url{https://github.com/TommasoBelluzzo/PyDTMC}}
+  howpublished = {\url{https://github.com/TommasoBelluzzo/PyDTMC}},
+  url={https://github.com/TommasoBelluzzo/PyDTMC}
   }
 
 @misc{simple-markov,
   title = {simple-markov},
-  howpublished = {\url{https://github.com/Mandragorian/simple-markov}}
+  howpublished = {\url{https://github.com/Mandragorian/simple-markov}},
+  url={https://github.com/Mandragorian/simple-markov}
   }
 
 @misc{mchmm,
   title = {mchmm},
-  howpublished = {\url{https://github.com/maximtrp/mchmm}}
+  howpublished = {\url{https://github.com/maximtrp/mchmm}},
+  url={https://github.com/maximtrp/mchmm}
   }
 
 @article{behnel2010cython,
@@ -30,7 +34,8 @@
   number={2},
   pages={31--39},
   year={2010},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/MCSE.2010.118},
 }
 
 
@@ -42,7 +47,8 @@
   number={7825},
   pages={357--362},
   year={2020},
-  publisher={Nature Publishing Group UK London}
+  publisher={Nature Publishing Group UK London},
+  doi={10.1038/s41586-020-2649-2}
 }
 
 
@@ -54,7 +60,8 @@
   number={3},
   pages={261--272},
   year={2020},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={10.1038/s41592-019-0686-2}
 }
 
 
@@ -72,7 +79,8 @@
   author={Lam, Siu Kwan and Pitrou, Antoine and Seibert, Stanley},
   booktitle={Proceedings of the Second Workshop on the LLVM Compiler Infrastructure in HPC},
   pages={1--6},
-  year={2015}
+  year={2015},
+  doi={10.1145/2833157.2833162}
 }
 
 
@@ -84,7 +92,8 @@
   number={3},
   pages={387--405},
   year={1995},
-  publisher={SIAM}
+  publisher={SIAM},
+  doi={10.1137/1037083}
 }
 
 @incollection{hamilton2010regime,
@@ -93,5 +102,6 @@
   booktitle={Macroeconometrics and time series analysis},
   pages={202--209},
   year={2010},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1057/9780230280830_23},
 }
