+++ JOSS/paper.bib
@@ -4,7 +4,7 @@
   booktitle={Proceedings of the IEEE Conference on Computer Vision and Pattern Recognition Workshops},
   pages={2180--2187},
   year={2018},
-  url={https://openaccess.thecvf.com/content_cvpr_2018_workshops/papers/w42/Awiszus_Markov_Chain_Neural_CVPR_2018_paper.pdf}
+  doi={10.1109/CVPRW.2018.00293}
 }
 
 
