+++ JOSS/paper.bib
@@ -70,7 +70,8 @@
   author={Paszke, Adam and Gross, Sam and Massa, Francisco and Lerer, Adam and Bradbury, James and Chanan, Gregory and Killeen, Trevor and Lin, Zeming and Gimelshein, Natalia and Antiga, Luca and others},
   journal={Advances in neural information processing systems},
   volume={32},
-  year={2019}
+  year={2019},
+  doi={10.5555/3454287.3455008}
 }
 
 
