+++ JOSS/paper.bib
@@ -65,13 +65,13 @@
 }
 
 
-@article{paszke2019pytorch,
-  title={Pytorch: An imperative style, high-performance deep learning library},
-  author={Paszke, Adam and Gross, Sam and Massa, Francisco and Lerer, Adam and Bradbury, James and Chanan, Gregory and Killeen, Trevor and Lin, Zeming and Gimelshein, Natalia and Antiga, Luca and others},
-  journal={Advances in neural information processing systems},
-  volume={32},
-  year={2019},
-  doi={10.5555/3454287.3455008}
+@inproceedings{ansel2024pytorch,
+  title={Pytorch 2: Faster machine learning through dynamic python bytecode transformation and graph compilation},
+  author={Ansel, Jason and Yang, Edward and He, Horace and Gimelshein, Natalia and Jain, Animesh and Voznesensky, Michael and Bao, Bin and Bell, Peter and Berard, David and Burovski, Evgeni and others},
+  booktitle={Proceedings of the 29th ACM International Conference on Architectural Support for Programming Languages and Operating Systems, Volume 2},
+  pages={929--947},
+  year={2024},
+  doi={10.1145/3620665.3640366}
 }
 
 
