### 2024-08-12 #6880 @aadya940
@mooniean I wasn't able to find the other DOIs, these likely don't exist. They're mostly just github repos and not proper publications.
