+++ .github/workflows/release.yml
@@ -40,3 +40,32 @@ jobs:
             pattern: cibw-*
             path: dist
             merge-multiple: true
+  
+  publish:
+    name: Publish to PyPI
+    runs-on: ubuntu-latest
+    needs: download
+
+    steps:
+      - uses: actions/checkout@v4
+
+      - name: Download built wheels
+        uses: actions/download-artifact@v4
+        with:
+          pattern: cibw-*
+          path: dist
+          merge-multiple: true
+
+      - name: Set up Python
+        uses: actions/setup-python@v4
+        with:
+          python-version: "3.12"  # Use appropriate Python version
+
+      - name: Install Twine
+        run: python -m pip install --upgrade twine
+
+      - name: Publish package to PyPI
+        env:
+          TWINE_USERNAME: ${{ secrets.TWINE_USERNAME }}
+          TWINE_PASSWORD: ${{ secrets.TWINE_PASSWORD }}
+        run: python -m twine upload dist/*
