### 2024-08-15 #6880 @mooniean
Hi @aadya940! Can you edit the zenodo and make sure the title and the author list match the paper title and author list?
