### 2024-08-15 4785366
update torch ref in JOSS paper
