### 2024-08-15 #6880 @crvernon
:wave: @aadya940 - I need to you to fix the following before I move to accept this for publication:

- [ ] The title of the Zenodo archive should match the title of the paper exactly.  No need to conduct another release, you can just edit the metadata in the Zenodo archive.
- [ ] The license you used in your repository should match the license selected in the Zenodo archive.  Please edit the metadata in the Zenodo archive to use the same license as your code.

- [ ] The title of the paper has mixed case capitalization.  Please change to "ChainoPy: A Python Library for Discrete Time Markov Chain Based Stochastic Analysis"
- [ ] LINE 37:  Missing comma after "typing" in "...dynamic typing etc."
- [ ] LINE 39:  "...chain, calculating properties."  should be "...chain, and calculating properties."
- [ ] LINE 50: "numpy" should be written as "NumPy" in this context
- [ ] LINE 60: "KL" is used here as an acronym before it has been defined as "Kullback–Leibler"
- [ ] LINES 62 through 112 should be removed.  They are not needed in the publication.
- [ ] LINE 114:  the "D" in "Documentation" should not be capitalized here.  Not should the "T" or "B" in testing and benchmarking
- [ ] LINE 123:  the "H" in histogram should not be capitalized in this context
- [ ] Ensure that your references have the proper capitalization where needed.  For example in line 131, the "p" in "python" should be capitalized.  You can maintain capitalization in your references in the bib file using curly brackets around the characters you wish to maintain for formatting of.
- [ ] Citations to those GitHub repos should have a date of retrieval or software release data for the version you are citing.  A bib format like the following may work:

```
@misc{yourid,
  title = {title of the package},
  author = {{Package's Development Team}},
  year = {four digit year},
  url = {url to the github repository or docs site},
}
```
