### 2024-08-15 2c43773
some fixes in JOSS paper
