+++ JOSS/paper.bib
@@ -10,20 +10,26 @@
 
 @misc{pydtmc,
   title = {PyDTMC},
+  author = {Tommaso Belluzzo},
   howpublished = {\url{https://github.com/TommasoBelluzzo/PyDTMC}},
-  url={https://github.com/TommasoBelluzzo/PyDTMC}
+  url={https://github.com/TommasoBelluzzo/PyDTMC},
+  year = {2024},
   }
 
 @misc{simple-markov,
   title = {simple-markov},
+  author = {Vasilis Charisopoulos and Kostis Andrikopoulos},
   howpublished = {\url{https://github.com/Mandragorian/simple-markov}},
-  url={https://github.com/Mandragorian/simple-markov}
+  url={https://github.com/Mandragorian/simple-markov},
+  year={2016},
   }
 
 @misc{mchmm,
   title = {mchmm},
+  author = {Maksim Terpilovskii},
   howpublished = {\url{https://github.com/maximtrp/mchmm}},
-  url={https://github.com/maximtrp/mchmm}
+  url={https://github.com/maximtrp/mchmm},
+  year={2021}
   }
 
 @article{behnel2010cython,
@@ -66,7 +72,7 @@
 
 
 @inproceedings{ansel2024pytorch,
-  title={Pytorch 2: Faster machine learning through dynamic python bytecode transformation and graph compilation},
+  title={Pytorch 2: Faster machine learning through dynamic Python bytecode transformation and graph compilation},
   author={Ansel, Jason and Yang, Edward and He, Horace and Gimelshein, Natalia and Jain, Animesh and Voznesensky, Michael and Bao, Bin and Bell, Peter and Berard, David and Burovski, Evgeni and others},
   booktitle={Proceedings of the 29th ACM International Conference on Architectural Support for Programming Languages and Operating Systems, Volume 2},
   pages={929--947},
@@ -76,7 +82,7 @@
 
 
 @inproceedings{lam2015numba,
-  title={Numba: A llvm-based python jit compiler},
+  title={Numba: A llvm-based Python jit compiler},
   author={Lam, Siu Kwan and Pitrou, Antoine and Seibert, Stanley},
   booktitle={Proceedings of the Second Workshop on the LLVM Compiler Infrastructure in HPC},
   pages={1--6},
