+++ JOSS/paper.md
@@ -36,62 +36,10 @@ We do the following key optimizations:
 
 The `MarkovChainNeuralNetwork` implementation defines a neural network model, using PyTorch [@ansel2024pytorch] for simulating Markov chain behavior. It takes a Markov chain object and the number of layers as input, with each layer being a linear layer. The model's forward method computes the output probabilities for the next state. The model is trained using stochastic gradient descent (SGD) with a learning rate scheduler. Finally, the model's performance is evaluated using the Kullback–Leibler divergence between the original Markov chain's transition probabilities and those estimated from the simulated walks.
 
-API of the library:
-
-    - chainopy.MarkovChain(transition-matrix: ndarray, states: list)
-            
-            Public Methods
-            -------------- 
-            
-            - fit()
-            - simulate()
-            - predict()
-            - adjacency_matrix()
-            - nstep_distribution()
-            - is_ergodic()
-            - is_symmetric()
-            - stationary_dist()
-            - is_absorbing()
-            - is_aperiodic()
-            - period()
-            - is_irreducible()
-            - is_transient(state)
-            - is_recurrent(state)
-            - fundamental_matrix()
-            - absorption_probabilities()
-            - expected_time_to_absorption()
-            - expected_number_of_visits()
-            - expected_hitting_time(state)
-            - visualize_transition_matrix()
-            - visualize_chain()
-            - save_model()
-            - load_model()
-            - marginal_dist()
-            - fit_from_file()
-
-    - chainopy.MarkovChainNeuralNetwork(chainopy.MarkovChain, num_layers)
-
-            Public Methods
-            --------------
-
-            - train_model()
-            - get_weights()
-            - simulate_random_walk()
-
-    - chainopy.MarkovSwitchingModel()
-
-            Public Methods
-            --------------
-
-            - fit()
-            - predict()
-            - evaluate()
-    
-    - chainopy.divergance_analysis(MarkovChain, MarkovChainNeuralNetwork)
 
 ### Documentation, Testing and Benchmarking
 
-For Documentation we use Sphinx. For Testing and Benchmarking we use the Pytest and PyDTMC [@pydtmc] packages. 
+For documentation we use Sphinx. For yesting and benchmarking we use the Pytest and PyDTMC [@pydtmc] packages. 
 
 The results are as follows:
 
@@ -146,7 +94,7 @@ The results are as follows:
 |                        | 5000    | 209 ms        | 16.4 ms          | 285 ms      | 27.6ms         |
 
 Apart from this, we test the `MarkovChainNeuralNetworks` by training them and comparing random walks between
-the original `MarkovChain` object and those generated by `MarkovChainNeuralNetworks` through a Histogram.
+the original `MarkovChain` object and those generated by `MarkovChainNeuralNetworks` through a histogram.
 
 ## Conclusion
 
