+++ JOSS/paper.bib
@@ -1,5 +1,5 @@
 @inproceedings{awiszus2018markov,
-  title={Markov chain neural networks},
+  title={Markov Chain neural networks},
   author={Awiszus, Maren and Rosenhahn, Bodo},
   booktitle={Proceedings of the IEEE Conference on Computer Vision and Pattern Recognition Workshops},
   pages={2180--2187},
@@ -72,7 +72,7 @@
 
 
 @inproceedings{ansel2024pytorch,
-  title={Pytorch 2: Faster machine learning through dynamic Python bytecode transformation and graph compilation},
+  title={PyTorch 2: Faster machine learning through dynamic Python bytecode transformation and graph compilation},
   author={Ansel, Jason and Yang, Edward and He, Horace and Gimelshein, Natalia and Jain, Animesh and Voznesensky, Michael and Bao, Bin and Bell, Peter and Berard, David and Burovski, Evgeni and others},
   booktitle={Proceedings of the 29th ACM International Conference on Architectural Support for Programming Languages and Operating Systems, Volume 2},
   pages={929--947},
@@ -82,7 +82,7 @@
 
 
 @inproceedings{lam2015numba,
-  title={Numba: A llvm-based Python jit compiler},
+  title={Numba: A LLVM-Based Python jit compiler},
   author={Lam, Siu Kwan and Pitrou, Antoine and Seibert, Stanley},
   booktitle={Proceedings of the Second Workshop on the LLVM Compiler Infrastructure in HPC},
   pages={1--6},
@@ -92,9 +92,9 @@
 
 
 @article{rosenthal1995convergence,
-  title={Convergence rates for Markov chains},
+  title={Convergence rates for Markov Chains},
   author={Rosenthal, Jeffrey S},
-  journal={Siam Review},
+  journal={SIAM Review},
   volume={37},
   number={3},
   pages={387--405},
