### 2024-08-15 #6880 @aadya940
@crvernon I've fixed the issues now, Does it look fine to you now?
