+++ JOSS/paper.md
@@ -39,7 +39,7 @@ The `MarkovChainNeuralNetwork` implementation defines a neural network model, us
 
 ### Documentation, Testing and Benchmarking
 
-For documentation we use Sphinx. For yesting and benchmarking we use the Pytest and PyDTMC [@pydtmc] packages. 
+For documentation we use Sphinx. For testing and benchmarking we use the Pytest and PyDTMC [@pydtmc] packages. 
 
 The results are as follows:
 
@@ -98,7 +98,6 @@ the original `MarkovChain` object and those generated by `MarkovChainNeuralNetwo
 
 ## Conclusion
 
-In conclusion, ChainoPy offers a Python library for discrete-time Markov Chains and includes features for Markov Chain Neural Networks, providing a useful tool for researchers and practitioners in stochastic analysis with efficient 
-performance.
+In conclusion, ChainoPy offers a Python library for discrete-time Markov Chains and includes features for Markov Chain Neural Networks, providing a useful tool for researchers and practitioners in stochastic analysis with efficient performance.
 
 ## References
